% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{adv_cut}
\alias{adv_cut}
\title{Cut numeric variable into groups (bins) with advanced options}
\usage{
adv_cut(
  x,
  ncuts = NULL,
  eq_groups = FALSE,
  cuts = NULL,
  num_labels = FALSE,
  labels = NULL,
  include_oob = TRUE,
  labels_at_means = TRUE,
  label_pairs_format = "[\%.2f, \%.2f]",
  ...
)
}
\arguments{
\item{x}{vector of numeric values to cut into groups}

\item{ncuts}{number of cuts (default: NULL)}

\item{eq_groups}{should the groups be equal (default: FALSE)}

\item{cuts}{where to put the cuts (default: NULL), not used if ncuts is used}

\item{num_labels}{should the labels be transformed into numbers (default: FALSE)}

\item{labels}{a vector of labels to use for the groups (default: NULL)}

\item{include_oob}{include values outside of the boundaries provided in `cuts` (default: TRUE)}

\item{labels_at_means}{should labels be created as means between cuts (T) or as pairs of cuts (F)}

\item{label_pairs_format}{formatting string to use when labels are generated from pairs of cuts (default: [\%.2f, \%.2f])}

\item{...}{other parameters passed to `base::cut`

If `ncuts` is used, then the variable is cut into N cuts either of equal group size (eq_groups = TRUE) or equally distant from each other (eq_groups = FALSE). If `labels` are not provided, they are generated as  means between cuts if labels_at_means is T.}
}
\value{
a vector of group labels the same length as the original value vector
}
\description{
Cut numeric variable into groups (bins) with advanced options
}
\examples{
set.seed(1)
x <- sample(1:100, 20)
sort(x)

adv_cut(x, ncuts = 5)
adv_cut(x, ncuts = 5, eq_groups = TRUE)
adv_cut(x, ncuts = 5, eq_groups = TRUE, num_labels = TRUE)
adv_cut(x, ncuts = 5, eq_groups = TRUE, labels_at_means = FALSE)
adv_cut(x, cuts = seq(0, 100, by = 20))
adv_cut(x, cuts = seq(0, 100, by = 20), labels_at_means = FALSE)
adv_cut(x, cuts = seq(0, 100, by = 20), labels_at_means = FALSE, label_pairs_format = "[\%i, \%i]")
}
