% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{aggr2}
\alias{aggr2}
\title{Double aggregation}
\usage{
aggr2(x, by, fun, ...)
}
\arguments{
\item{x}{value to aggregate}

\item{by}{vector to aggregate by (e.g., ID of participant)}

\item{fun}{function to apply}

\item{...}{additional parameters passed to fun}
}
\value{
value aggregated first by specified vector and then aggregated again
}
\description{
Aggregates value twice providing mean of means, SD of SDs, etc.
}
\examples{
x<-rnorm(100)
id<-rep(1:10, each=10)

aggregate(x~id, FUN=mean)
aggr2(x, id, mean)
}

