% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_functions.R
\name{apa.afex_aov}
\alias{apa.afex_aov}
\title{Describe afex ANOVA results}
\usage{
\method{apa}{afex_aov}(
  obj,
  term,
  include_eta = TRUE,
  eta_digits = 2,
  f_digits = 2,
  df_digits = 0,
  append_to_table = FALSE,
  ...
)
}
\arguments{
\item{obj}{afex ANOVA object}

\item{term}{model term to describe (a string with the term name or its sequential number)}

\item{include_eta}{add generalized eta^2 for the model (default: TRUE)}

\item{eta_digits}{number of digits to use for eta^2 (default: 2)}

\item{f_digits}{number of digits to use for F (default: 2)}

\item{df_digits}{number of digits to use for df (default: 0)}

\item{append_to_table}{should the results be added to the afex ANOVA table (default: FALSE)}

\item{...}{other parameters passed to \link{format_results}}
}
\value{
string with formatted results
}
\description{
Provides formatted string like \emph{F}(num.Df, den.Df) = ..., \emph{p} ..., eta2 = ... based on \link[afex:aov_car]{afex::aov_ez}, \link[afex:aov_car]{afex::aov_car}, or \link[afex:aov_car]{afex::aov_4} results.
}
\examples{
if (requireNamespace("afex", quietly = TRUE)) {
  data(stroop, package = "afex")

  afex_res <- afex::aov_ez(
    id = "pno",
    dv = "rt",
    within = "condition",
    data = stroop[!is.na(stroop$rt), ],
    fun_aggregate = mean
  )

  apa(afex_res, "condition")
  apa(afex_res, "condition", eta_digits = 3)
  apa(afex_res, "condition", include_eta = FALSE)
}
}
