% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apa.chi}
\alias{apa.chi}
\title{Describe $chi^2$ results}
\usage{
\method{apa}{chi}(tbl, v = TRUE, addN = TRUE, ...)
}
\arguments{
\item{tbl}{\itemize{
\item MxN table on which to compute chi^2 (if not a table, will try to make a table out of it)
}}

\item{v}{\itemize{
\item add Cramer's V (default: T)
}}

\item{addN}{\itemize{
\item add N (default: T)
}}

\item{...}{other parameters passed to \link{format.results}}
}
\value{
result
}
\description{
Describe Pearson $chi^2$ results. Note that the function takes table as an input rather than $chi^2$ object - this is necessary to get Cramer's V.
}
\examples{

## From chisq.test help page
## ---------------------
## From Agresti(2007) p.39
M <- as.table(rbind(c(762, 327, 468), c(484, 239, 477)))
dimnames(M) <- list(
  gender = c("FALSE", "M"),
  party = c("Democrat", "Independent", "Republican")
)
(Xsq <- chisq.test(M)) # Prints test summary
## ----------------------
apa.chi(M) # Note that apa.chi should be used for the table, not the chi2

}
