% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apa.emmeans}
\alias{apa.emmeans}
\title{Describe contrasts created by emmeans}
\usage{
\method{apa}{emmeans}(obj, term, dtype = "B", df = FALSE, ...)
}
\arguments{
\item{obj}{summary object from emmeans::contrast}

\item{term}{contrast number(s)}

\item{dtype}{description type, "t", "B", or any other letter}

\item{df}{include DF in t-test description (default: False)}

\item{...}{other parameters passed to \link{format.results}}
}
\value{
string with formatted results
}
\description{
Describe contrasts created by emmeans
}
\examples{

require(emmeans)
warp.lm <- lm(breaks ~ wool * tension, data = warpbreaks)
warp.lsm <- emmeans::emmeans(warp.lm, ~ tension | wool)
(sum_contr <- summary(contrast(warp.lsm, "trt.vs.ctrl")))
apa(sum_contr, 1)
apa(sum_contr, 3)
apa(sum_contr, 3, dtype = "t")
apa(sum_contr, 3, dtype = "c")
apa(sum_contr, 3, dtype = "c", df = TRUE)
}
