% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apa.ezstats}
\alias{apa.ezstats}
\title{Describe ezStats results}
\usage{
\method{apa}{ezstats}(ezstats_res, term = 1, ...)
}
\arguments{
\item{ezstats_res}{data.frame returned by ezStats}

\item{term}{model term to describe (a string with the term name or its sequential number)}

\item{...}{other parameters passed to apa.mean.sd

If a string is used for a term and there is more than one factor, "X:Y:Z" format is assumed (value in first column : value in the second, and so on). If no term is supplied, the first row is used. So if you need to select a term based on two or more variables, you can also just filter ezStats result beforehand (or use a row number as a term).}
}
\value{
string with formatted results
}
\description{
Returns formatted string with mean and SD from ezStats object
}
\examples{
data(faces)
faces$uid <- factor(faces$uid)
faces$correct <- factor(faces$correct)
ezstats_res <- ez::ezStats(faces, dv = answerTime, wid = uid, within = .(stim_gender, correct))
ezstats_res
apa(ezstats_res, "M:0")
apa(ezstats_res, 2)
apa(ezstats_res, 2, digits = 1)

}
