% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{apa.htest}
\alias{apa.htest}
\title{Format t-test results in APA style}
\usage{
\method{apa}{htest}(obj, show.mean = FALSE, abs = FALSE, ...)
}
\arguments{
\item{obj}{An object from \link[stats:t.test]{stats::t.test}}

\item{show.mean}{Include mean value in results (useful for one-sample test)}

\item{abs}{Show the absolute value of t-statistic}

\item{...}{Other arguments passed to \link{format_results}}
}
\value{
A string with t-test value, degrees of freedom, p-value, and optionally the mean with CI
}
\description{
Format t-test results in APA style
}
\examples{
# One-sample t-test
t_res <- t.test(rnorm(20, mean = -10, sd = 2))
apa(t_res)
apa(t_res, show.mean = TRUE)

# Two-sample t-test
t_two <- t.test(rnorm(20), rnorm(20, mean = 0.8))
apa(t_two)
}
