% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apa.lht}
\alias{apa.lht}
\title{Describe linearHypothesis test results}
\usage{
\method{apa}{lht}(hyp, ...)
}
\arguments{
\item{hyp}{hypothesis from \link[car]{linearHypothesis}}

\item{...}{additional papameters passed to \link{format.results}}
}
\value{
results of \eqn{\chi^2} or \emph{F} test
}
\description{
Describe linearHypothesis test results
}
\examples{

mod.davis <- lm(weight ~ repwt, data = carData::Davis)

res <- car::linearHypothesis(mod.davis, c("(Intercept) = 0", "repwt = 1"))
apa(res)

res.chi <- car::linearHypothesis(mod.davis, c("(Intercept) = 0", "repwt = 1"), test = "Chisq")
apa(res.chi)
}
