% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_functions.R
\name{apa.linearHypothesis}
\alias{apa.linearHypothesis}
\title{Describe linearHypothesis test results}
\usage{
\method{apa}{linearHypothesis}(obj, ...)
}
\arguments{
\item{obj}{hypothesis from \link[car:linearHypothesis]{car::linearHypothesis}}

\item{...}{additional parameters passed to \link{format_results}}
}
\value{
results of \eqn{\chi^2} or \emph{F} test
}
\description{
Describe linearHypothesis test results
}
\examples{
if (requireNamespace("car", quietly = TRUE) &&
    requireNamespace("carData", quietly = TRUE)) {
  # Create a linear model
  mod.davis <- lm(weight ~ repwt, data = carData::Davis)
  
  # Test hypotheses
  res <- car::linearHypothesis(mod.davis, c("(Intercept) = 0", "repwt = 1"))
  apa(res)
  
  # Test using Chi-square
  res.chi <- car::linearHypothesis(mod.davis, c("(Intercept) = 0", "repwt = 1"), test = "Chisq")
  apa(res.chi)
}
}
