% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apa.lmer}
\alias{apa.lmer}
\title{Describe lmer results}
\usage{
\method{apa}{lmer}(fm, pv, digits = c(2, 2, 2), incl.rel = 0, dtype = "B", incl.p = TRUE)
}
\arguments{
\item{fm}{LMER model from lme4}

\item{pv}{p values table (from Anova)}

\item{digits}{number of digits in results (default: 2)}

\item{incl.rel}{include the relation sign for \emph{p}}

\item{dtype}{description type (B or t)}

\item{incl.p}{include p values}
}
\value{
result
}
\description{
This function is deprecated in favor of \link{apa.glm}
}
