% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{apa.lmerMod}
\alias{apa.lmerMod}
\alias{apa.lmer}
\title{Describe lmer results}
\usage{
\method{apa}{lmerMod}(obj, pv, digits = c(2, 2, 2), incl.rel = 0, dtype = "B", incl.p = TRUE)

\method{apa}{lmer}(obj, pv, digits = c(2, 2, 2), incl.rel = 0, dtype = "B", incl.p = TRUE)
}
\arguments{
\item{obj}{LMER model from \link[lme4:lmer]{lme4::lmer}}

\item{pv}{p values table (from Anova)}

\item{digits}{number of digits in results (default: 2)}

\item{incl.rel}{include the relation sign for \emph{p}}

\item{dtype}{description type (B or t)}

\item{incl.p}{include p values}
}
\value{
result
}
\description{
Describe lmer results
}
\examples{
if (requireNamespace("lme4", quietly = TRUE)) {
  # Fit a mixed-effects model
  fm <- lme4::lmer(Reaction ~ Days + (Days | Subject), lme4::sleepstudy)
  
  # Get approximate p-values
  if (requireNamespace("lmerTest", quietly = TRUE)) {
    pv <- list(fixed = data.frame("Pr(>|t|)" = c(0.001, 0.001)))
    rownames(pv$fixed) <- c("(Intercept)", "Days")
    
    # Format results
    apa(fm, pv)
    apa(fm, pv, digits = c(1, 1, 1))
    apa(fm, pv, dtype = "t")
  }
}
}
