% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apa.lmert}
\alias{apa.lmert}
\title{Describe lmerTest results}
\usage{
\method{apa}{lmert}(sfit, factor, dtype = "t", ...)
}
\arguments{
\item{sfit}{\emph{summary} object from lmerTest::lmer model}

\item{factor}{name or number of the factor that needs to be describe}

\item{dtype}{description type ("B"/"t")}

\item{...}{other parameters passed to \link{format.results}}
}
\value{
Formatted string
}
\description{
Note that this function uses \emph{summary} object from lmerTest::lmer model and not the model itself (see the example). Otherwise p-values will be computed during the call and everything will be very slow.
}
\examples{

fm <- lmerTest::lmer(Reaction ~ Days + (Days | Subject), lme4::sleepstudy)
fms <- summary(fm)
apa(fms, "Days")
apa(fms, "(Intercept)")
apa(fms, 2)
apa(fms, "Days", "B")

}
