% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apa.lmtaov}
\alias{apa.lmtaov}
\title{Describe lmerTest anova results}
\usage{
\method{apa}{lmtaov}(afit, term, f.digits = 2, ...)
}
\arguments{
\item{afit}{\itemize{
\item lmerTest anova results
}}

\item{term}{model term to describe (a string with the term name or its sequential number)}

\item{f.digits}{\itemize{
\item decimal digits for F value
}}

\item{...}{other parameters passed to \link{format.results}}
}
\value{
formatted string describing the results of anova
}
\description{
Describe lmerTest anova results
}
\examples{
data(faces)

fit <- lmerTest::lmer(answerTime ~ correct * stim_gender + (1 + correct * stim_gender | uid), faces)
afit <- anova(fit)
apa(afit, "correct:stim_gender")
}
