% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apa.mean.sd}
\alias{apa.mean.sd}
\title{Describe mean and SD}
\usage{
\method{apa}{mean.sd}(
  x = NULL,
  m = NULL,
  sd = NULL,
  digits = 2,
  dtype = "p",
  m_units = "",
  sd_units = "",
  ...
)
}
\arguments{
\item{x}{vector of values}

\item{m}{mean (not used if x is provided)}

\item{sd}{SD (not used if x is provided)}

\item{digits}{number of digits used for description (default: 2)}

\item{dtype}{"p" for SD in parentheses (default), "c" for SD after comma}

\item{m_units}{a character to add units (e.g., "°" for degrees) to the mean}

\item{sd_units}{a character to add units (e.g., "°" for degrees) to SD}

\item{...}{other arguments passed to \link{format.results}}
}
\value{
formatted string
}
\description{
Computes mean and SD for a vector (or just uses mean and sd if provided) and returns them as formatted string
}
\examples{

x <- rnorm(1000, 50, 25)
apa.mean.sd(x)
apa.mean.sd(x, 0)
apa.mean.sd(x, dtype = "c")
}
