% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{apa}
\alias{apa}
\title{Format statistical results in APA style}
\usage{
apa(obj, ...)
}
\arguments{
\item{obj}{An object to format in APA style}

\item{...}{Additional arguments passed to specific methods}
}
\value{
A formatted string in APA style
}
\description{
The \code{apa} function provides a convenient way to format statistical results
according to American Psychological Association (APA) style guidelines.
It dispatches to various S3 methods depending on the class of the input object.
}
\details{
This function supports various statistical objects directly:
\itemize{
\item \link{apa.ttest} for \link[stats:t.test]{stats::t.test} and \link{apa.r} for \link[stats:cor.test]{stats::cor.test} (both with class \code{htest})
\item \link{apa.aov} for \link[stats:aov]{stats::aov} and \link{apa.anova} for \link[car:Anova]{car::Anova}
\item \link{apa.glm} for \link[stats:glm]{stats::glm}, \link[stats:lm]{stats::lm}, \link[lme4:lmer]{lme4::lmer}, and \link{apa.lmert} for \link[lmerTest:lmer]{lmerTest::lmer}
\item \link{apa.brmsfit} for \link[brms:brm]{brms::brm} and \link{apa.BFBayesFactor} for \link[BayesFactor:anovaBF]{BayesFactor::anovaBF}
\item \link{apa.chisq.test} for \link[stats:chisq.test]{stats::chisq.test}
\item \link{apa.dip.test} for \link[diptest:dip.test]{diptest::dip.test}
\item \link{apa.roc.test} for \link[pROC:roc.test]{pROC::roc.test}
}

In addition, there are standalone utility functions for formatting:
\itemize{
\item \link{apa_mean_sd}: Format means and standard deviations
\item \link{apa_mean_conf}: Format means and confidence intervals
\item \link{apa_binom_mean_conf}: Format binomial means and confidence intervals
\item \link{apa_mean_and_t}: Compare means with t-tests and format results
\item \link{table_mean_conf}: Get formatted means and CIs (for tables)
\item \link{table_mean_conf_by}: Get formatted means and CIs by group (for tables)
}
}
\examples{
# t-test example
t_res <- t.test(rnorm(30), rnorm(30, 0.5))
apa(t_res)

# Correlation example
cor_res <- cor.test(mtcars$mpg, mtcars$wt)
apa(cor_res)
}
