% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apa.roc.diff}
\alias{apa.roc.diff}
\title{Describe differences between ROC curves}
\usage{
\method{apa}{roc.diff}(roc_diff)
}
\arguments{
\item{roc_diff}{a difference between the ROC curves}
}
\value{
result
}
\description{
Describe differences between ROC curves
}
