% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{apa.summary.merMod}
\alias{apa.summary.merMod}
\alias{apa.lmert}
\title{Describe lmerTest results}
\usage{
\method{apa}{summary.merMod}(obj, factor, dtype = "t", ...)

\method{apa}{lmert}(obj, factor, dtype = "t", ...)
}
\arguments{
\item{obj}{\emph{summary} object from \link[lmerTest:lmer]{lmerTest::lmer} model}

\item{factor}{name or number of the factor that needs to be described}

\item{dtype}{description type ("B"/"t")}

\item{...}{other parameters passed to \link{format_results}}
}
\value{
Formatted string
}
\description{
Note that this function uses \emph{summary} object from lmerTest::lmer model and not the model itself (see examples). Otherwise p-values will be computed during the call and everything will be very slow.
}
\examples{
if (requireNamespace("lmerTest", quietly = TRUE)) {
  # Fit a mixed-effects model with p-values
  fm <- lmerTest::lmer(Reaction ~ Days + (Days | Subject), lme4::sleepstudy)
  fms <- summary(fm)
  
  # Format results in APA style
  apa(fms, "Days")
  apa(fms, "(Intercept)")
  apa(fms, 2)
  apa(fms, "Days", "B")
}
}
