% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_functions.R
\name{apa.summary_emm}
\alias{apa.summary_emm}
\alias{apa.emmeans}
\title{Describe contrasts created by emmeans}
\usage{
\method{apa}{summary_emm}(obj, term, dtype = "B", df = FALSE, ...)

\method{apa}{emmeans}(obj, term, dtype = "B", df = FALSE, ...)
}
\arguments{
\item{obj}{summary object from \link[emmeans:contrast]{emmeans::contrast}}

\item{term}{contrast number(s)}

\item{dtype}{description type, "t", "B", or any other letter}

\item{df}{include DF in t-test description (default: False)}

\item{...}{other parameters passed to \link{format_results}}
}
\value{
string with formatted results
}
\description{
Describe contrasts created by emmeans
}
\examples{
if (requireNamespace("emmeans", quietly = TRUE)) {
  # Create a model
  warp.lm <- lm(breaks ~ wool * tension, data = warpbreaks)
  
  # Create estimated marginal means
  warp.emm <- emmeans::emmeans(warp.lm, ~ tension | wool)
  
  # Create contrasts
  sum_contr <- summary(emmeans::contrast(warp.emm, "trt.vs.ctrl"))
  
  # Format results in APA style
  apa(sum_contr, 1)
  apa(sum_contr, 3)
  apa(sum_contr, 3, dtype = "t")
  apa(sum_contr, 3, dtype = "c")
  apa(sum_contr, 3, dtype = "c", df = TRUE)
}
}
