% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{apa_binom_mean_conf}
\alias{apa_binom_mean_conf}
\title{Format binomial proportions and confidence intervals in APA style}
\usage{
apa_binom_mean_conf(x, digits = 2, ...)
}
\arguments{
\item{x}{A vector of zeros and ones}

\item{digits}{Number of digits in results}

\item{...}{Additional arguments passed to \link{format_results}}
}
\value{
A string with the mean and confidence interval in square brackets
}
\description{
Format binomial proportions and confidence intervals in APA style
}
\examples{
# Generate binomial data
set.seed(123)
x <- rbinom(500, 1, prob = 0.7)

# Format results in APA style
apa_binom_mean_conf(x)
}
