% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_functions.R
\name{apa_ezStats}
\alias{apa_ezStats}
\title{Describe ezStats results}
\usage{
apa_ezStats(obj, term = 1, ...)
}
\arguments{
\item{obj}{data.frame returned by \link[ez:ezStats]{ez::ezStats}}

\item{term}{model term to describe (a string with the term name or its sequential number)}

\item{...}{other parameters passed to \link{apa_mean_sd}

If a string is used for a term and there is more than one factor, "X:Y:Z" format is assumed (value in first column : value in the second, and so on). If no term is supplied, the first row is used. So if you need to select a term based on two or more variables, you can also just filter ezStats result beforehand (or use a row number as a term).}
}
\value{
string with formatted results
}
\description{
Returns formatted string with mean and SD from ezStats object
}
\examples{
if (requireNamespace("ez", quietly = TRUE) && requireNamespace("afex", quietly = TRUE)) {
  # Using the Stroop dataset from afex package
  data(stroop, package = "afex")
  # Get descriptive statistics

  ezstats_res <- ez::ezStats(
    data = stroop[!is.na(stroop$rt),],
    dv = rt,
    wid = pno,
    within = .(congruency)
  )
  
  # Format results in APA style
  apa_ezStats(ezstats_res, 1)
  apa_ezStats(ezstats_res, "incongruent")
  apa_ezStats(ezstats_res, 1, digits = 3)
}
}
