% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{apa_format_mean_conf}
\alias{apa_format_mean_conf}
\title{Format mean and confidence interval values into APA-style string}
\usage{
apa_format_mean_conf(
  mean_val,
  lower_ci,
  upper_ci,
  addCI = FALSE,
  digits = 2,
  ...
)
}
\arguments{
\item{mean_val}{The mean value}

\item{lower_ci}{Lower confidence interval bound}

\item{upper_ci}{Upper confidence interval bound}

\item{addCI}{Add "95\% CI =" prefix}

\item{digits}{Number of digits to use}

\item{...}{Other arguments passed to \link{format_results}}
}
\value{
A formatted string with mean and confidence intervals
}
\description{
Format mean and confidence interval values into APA-style string
}
\examples{
# Example usage
mean_val <- 5.67
lower_ci <- 4.56
upper_ci <- 6.78
# Format mean and CI in APA style
apa_format_mean_conf(mean_val, lower_ci, upper_ci)
# Format mean and CI with additional CI prefix
apa_format_mean_conf(mean_val, lower_ci, upper_ci, addCI = TRUE)
# Format mean and CI with custom number of digits
apa_format_mean_conf(mean_val, lower_ci, upper_ci, digits = 3)

}
