% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{apa_mean_conf}
\alias{apa_mean_conf}
\title{Format means and confidence intervals in APA style}
\usage{
apa_mean_conf(
  x,
  bootCI = TRUE,
  addCI = FALSE,
  digits = 2,
  transform.means = NULL,
  ...
)
}
\arguments{
\item{x}{A numeric vector}

\item{bootCI}{Use bootstrapped CI (TRUE) or normal approximation (FALSE)}

\item{addCI}{Add "95\% CI =" prefix}

\item{digits}{Number of digits to use}

\item{transform.means}{An optional function to transform the means and CI to another scale}

\item{...}{Other arguments passed to \link{format_results}}
}
\value{
A string with a mean followed by confidence intervals in square brackets
}
\description{
Format means and confidence intervals in APA style
}
\examples{
x <- runif(100, 0, 50)
apa_mean_conf(x)
apa_mean_conf(x, bootCI = FALSE)  # Normal approximation instead of bootstrap
}
