% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apastats-package.R
\docType{package}
\name{apastats-package}
\alias{apastats}
\alias{apastats-package}
\title{apastats: APA-style results}
\description{
Proper formatting of statistical results in accordance with American Psychological Association Manual, 6th ed., and some convenient plotting functions.
}
\keyword{internal}
