% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apastats-package.R
\docType{package}
\name{apastats2-package}
\alias{apastats2}
\alias{apastats2-package}
\title{apastats2: Statistical Results in 'APA' Style}
\description{
Proper formatting of statistical results in accordance with the American Psychological Association ('APA') Manual (sixth edition), along with convenient helpers for summaries, confidence intervals, and plotting. Development repository: \url{https://github.com/achetverikov/APAstats}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/achetverikov/APAstats}
  \item Report bugs at \url{https://github.com/achetverikov/APAstats/issues}
}

}
\author{
\strong{Maintainer}: Andrey Chetverikov \email{andrey.chetverikov@uib.no} (\href{https://orcid.org/0000-0003-2767-6310}{ORCID})

}
\keyword{internal}
