% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{base.breaks.x}
\alias{base.breaks.x}
\title{Tufte-like breaks for X axis}
\usage{
base.breaks.x(x, addSegment = TRUE, ...)
}
\arguments{
\item{x}{vector of numbers to create breaks from}

\item{addSegment}{should we add a line to the scale? (T/F)}
}
\value{
scale_x_continuous with pretty breaks
}
\description{
Tufte-like breaks for X axis
}
