% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{base.breaks.y}
\alias{base.breaks.y}
\title{Title}
\usage{
base.breaks.y(x, expand = c(0, 0), ...)
}
\arguments{
\item{x}{}

\item{expand}{}

\item{...}{}
}
\value{
scale_y_continuous with pretty breaks
}
\description{
Title
}

