% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{base_breaks}
\alias{base_breaks}
\alias{base_breaks_x}
\alias{base_breaks_y}
\title{Tufte-like breaks for axes}
\usage{
base_breaks(x, scale = "x", addSegment = TRUE, ...)

base_breaks_x(x, addSegment = TRUE, ...)

base_breaks_y(x, addSegment = TRUE, ...)
}
\arguments{
\item{x}{vector of numbers to create breaks from}

\item{scale}{scale (x or y) to create breaks for}

\item{addSegment}{should we add a line to the scale? (T/F)}

\item{...}{other parameters passed to scale_(x or y)_continuous}
}
\value{
scale_(x or y)_continuous with pretty breaks and accompanying geom_segment if addSegment == TRUE
}
\description{
Tufte-like breaks for axes
}
\section{Functions}{
\itemize{
\item \code{base_breaks_x()}: Tufte-like breaks for X axis

\item \code{base_breaks_y()}: Tufte-like breaks for Y axis

}}
\examples{
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  theme_minimal()

p
p + base_breaks(mtcars$wt, scale = "x") +
  base_breaks(mtcars$mpg, scale = "y")
}
