% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{binom.ci}
\alias{binom.ci}
\title{Binomial confidence interval as a vector}
\usage{
binom.ci(x)
}
\arguments{
\item{x}{a vector of 0 and 1}
}
\value{
a vector of mean, lower CI, upper CI, and length of x
}
\description{
Binomial confidence interval as a vector
}
\examples{
binom.ci(rbinom(500, 1, prob=0.7))
}

