% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.aov}
\alias{describe.aov}
\title{Describe aov results}
\usage{
describe.aov(fit, term, type = 2, ...)
}
\arguments{
\item{fit}{}

\item{term}{}

\item{type}{}

\item{...}{}
}
\value{
result
}
\description{
Describe aov results
}

