% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.bf}
\alias{describe.bf}
\title{Describe BayesFactor results}
\usage{
describe.bf(bf, digits = 2, ...)
}
\arguments{
\item{bf}{an object of \code{BFBayesFactor} class}

\item{digits}{number of digits to use}

\item{...}{other parameters passed to format.results}
}
\value{
string describing the result
}
\description{
Describe BayesFactor results
}
\examples{

require('BayesFactor')
data(puzzles)
bfs <- anovaBF(RT ~ shape*color + ID, data = puzzles, whichRandom = "ID", progress=FALSE)
describe.bf(bfs[1])
describe.bf(bfs[1]/bfs[2])

}

