% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.bimod.test}
\alias{describe.bimod.test}
\title{Describe bimodality test results}
\usage{
describe.bimod.test(x, start_vec = NA, ...)
}
\arguments{
\item{x}{}

\item{start_vec}{}

\item{...}{}
}
\value{
result
}
\description{
Describe bimodality test results
}

