% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.binom.mean.conf}
\alias{describe.binom.mean.conf}
\title{Describe mean and confidence intervals for binomial variable}
\usage{
describe.binom.mean.conf(x, digits = 2)
}
\arguments{
\item{x}{}

\item{digits}{}
}
\value{
result
}
\description{
Describe mean and confidence intervals for binomial variable
}

