% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.brm}
\alias{describe.brm}
\title{Describe brms model results}
\usage{
describe.brm(
  mod,
  term,
  trans = NULL,
  digits = 2,
  eff.size = FALSE,
  eff.size.type = "r",
  nsamples = 100,
  ci.type = "HPDI",
  ...
)
}
\arguments{
\item{mod}{model object}

\item{term}{model term to describe}

\item{trans}{an optional function to transform the results to another scale (default: NULL)}

\item{digits}{number of digits in the output}

\item{eff.size}{string describing how to compute an effect size (currently, either 'fe_to_all', 'part_fe', or 'part_fe_re')}

\item{eff.size.type}{type of the effect size ('r' or 'r2')}

\item{nsamples}{number of samples to use for the effect size computations}

\item{ci.type}{type of intervals to use (currently, all that is not HPDI is treated as ETI using `bayestestR::eti`)}

\item{...}{other parameters passed to \link{format.results}}
}
\value{
string describing the result
}
\description{
Describe brms model results
}
\examples{
\dontrun{
require(brms)
x <- rnorm(500, sd = 6)
y <- 4 * x + rnorm(500)
fit <- brm(y ~ x, data = data.frame(x, y), chains = 1, iter = 500)
describe.brm(fit, "x")
# convert x to another scale for output
describe.brm(fit, "x", trans = function(val) val + 5)
}
}
