% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.chi}
\alias{describe.chi}
\title{Describe $chi^2$ results}
\usage{
describe.chi(tbl, v = T, addN = T, ...)
}
\arguments{
\item{...}{}
}
\value{
result
}
\description{
Describe $chi^2$ results
}

