% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.dip.test}
\alias{describe.dip.test}
\title{Describe Hartigans' dip test results}
\usage{
describe.dip.test(x, ...)
}
\arguments{
\item{x}{Hartigans' dip test results}

\item{...}{other parameters passed to \link{format.results}}
}
\value{
result
}
\description{
Describe Hartigans' dip test results
}
