% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.ezanova}
\alias{describe.ezanova}
\title{Describe ezANOVA results}
\usage{
describe.ezanova(ezfit, term, include_eta = T, spher_corr = T, ...)
}
\arguments{
\item{ezfit}{}

\item{term}{}

\item{include_eta}{}

\item{spher_corr}{}

\item{...}{}
}
\value{
result
}
\description{
Describe ezANOVA results
}

