% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.ezstats}
\alias{describe.ezstats}
\title{Describe ezStats results}
\usage{
describe.ezstats(ezstats_res, term = 1, ...)
}
\arguments{
\item{term}{model term to describe (a string with the term name or its sequential number)}

\item{...}{other parameters passed to describe.mean.sd

If a string is used for a term and there is more than one factor, "X:Y:Z" format is assumed (value in first column : value in the second, and so on). If no term is supplied, the first row is used. So if you need to select a term based on two or more variables, you can also just filter ezStats result beforehand (or use a row number as a term).}

\item{ez_stats}{data.frame returned by ezStats}
}
\value{
string with formatted results
}
\description{
Returns formatted string with mean and SD from ezStats object
}
\examples{
library(ez)
data(faces)
ezstats_res <- ezStats(faces, dv = answerTime, wid = uid, within = .(stim_gender,correct))
ezstats_res
describe.ezstats(ezstats_res, 'M:0')
describe.ezstats(ezstats_res, 2)
describe.ezstats(ezstats_res, 2, digits=1)

}
