% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.glm}
\alias{describe.glm}
\title{Describe regression model (GLM, GLMer, lm, lm.circular, ...)}
\usage{
describe.glm(fit, term = NULL, short = 1, f.digits = 2, test_df = F,
  p_as_number = F, term_pattern = NULL, ...)
}
\arguments{
\item{fit}{}

\item{term}{}

\item{short}{}

\item{f.digits}{}

\item{test_df}{}

\item{...}{}
}
\value{
result
}
\description{
Describe regression model (GLM, GLMer, lm, lm.circular, ...)
}

