% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.lht}
\alias{describe.lht}
\title{Describe linearHypothesis test results}
\usage{
describe.lht(hyp, ...)
}
\arguments{
\item{hyp}{hypothesis from \link[car]{linearHypothesis}}

\item{...}{additional papameters passed to \link{format.results}}
}
\value{
results of \eqn{\chi^2} or _F_ test
}
\description{
Describe linearHypothesis test results
}
\examples{

library(car)
mod.davis <- lm(weight ~ repwt, data=Davis)

res <- linearHypothesis(mod.davis, c("(Intercept) = 0", "repwt = 1"))
describe.lht(res)

res.chi <- linearHypothesis(mod.davis, c("(Intercept) = 0", "repwt = 1"), test = "Chisq")
describe.lht(res.chi)
}
