% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.lmert}
\alias{describe.lmert}
\title{Describe lmerTest results}
\usage{
describe.lmert(sfit, factor, dtype = "t", ...)
}
\arguments{
\item{sfit}{}

\item{factor}{}

\item{dtype}{}

\item{...}{}
}
\value{
result
}
\description{
Describe lmerTest results
}

