% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.lmtaov}
\alias{describe.lmtaov}
\title{Describe lmerTest anova results}
\usage{
describe.lmtaov(afit, term, f.digits = 2, ...)
}
\arguments{
\item{afit}{- lmerTest anova results}

\item{term}{- term name}

\item{f.digits}{- decimal digits for F}

\item{...}{}
}
\value{
formatted string describing the results of anova
}
\description{
Describe lmerTest anova results
}
\examples{
load.libs(c('lme4','lmerTest'))
data(faces)

fit <- lmer(answerTime~correct*stim_gender+(1+correct*stim_gender|uid), faces)
afit <- anova(fit)
describe.lmtaov(afit, 'correct:stim_gender')
}

