% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.lsmeans}
\alias{describe.lsmeans}
\title{Describe contrasts created by lsmeans}
\usage{
describe.lsmeans(obj, term, dtype = "B", df = FALSE, ...)
}
\arguments{
\item{obj}{summary object from lsmeans::contrast}

\item{term}{contrast number}

\item{dtype}{description type, "t", "B", or any other letter}

\item{df}{include DF in t-test description (default: False)}

\item{...}{other parameters passed to \link{format.results}}
}
\value{
string with formatted results
}
\description{
Describe contrasts created by lsmeans
}
\examples{

requireNamespace('lsmeans')
warp.lm <- lm(breaks ~ wool*tension, data = warpbreaks)
warp.lsm <- lsmeans::lsmeans(warp.lm, ~ tension | wool)
(sum_contr<-summary(lsmeans::contrast(warp.lsm, 'trt.vs.ctrl')))
describe.lsmeans(sum_contr, 1)
describe.lsmeans(sum_contr, 3)
describe.lsmeans(sum_contr, 3, dtype='t')
describe.lsmeans(sum_contr, 3, dtype='c')
describe.lsmeans(sum_contr, 3, dtype='c', df=TRUE)
}
