% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.mean.and.t}
\alias{describe.mean.and.t}
\title{Describe t-test with means}
\usage{
describe.mean.and.t(x, by, which.mean = 1, digits = 2, paired = F, ...)
}
\arguments{
\item{x}{}

\item{by}{}

\item{which.mean}{}

\item{digits}{}

\item{paired}{}

\item{...}{}
}
\value{
result
}
\description{
Describe t-test with means
}

