% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.mean.sd}
\alias{describe.mean.sd}
\title{Describe mean and SD}
\usage{
describe.mean.sd(x, digits = 2, ...)
}
\arguments{
\item{x}{}

\item{digits}{}

\item{...}{}
}
\value{
result
}
\description{
Describe mean and SD
}

