% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.roc.diff}
\alias{describe.roc.diff}
\title{Describe differences between ROC curves}
\usage{
describe.roc.diff(roc_diff)
}
\arguments{
\item{roc_diff}{}
}
\value{
result
}
\description{
Describe differences between ROC curves
}

