% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.ttest}
\alias{describe.ttest}
\title{Describe t-test results}
\usage{
describe.ttest(t, show.mean = F, abs = F, ...)
}
\arguments{
\item{t}{an object from \link[stats]{t.test}}

\item{show.mean}{include mean value in results (useful for one-sample test)}

\item{abs}{should we show the sign of t-test}

\item{...}{other arguments passed to \link{format.results}}
}
\value{
A string with t-test value, degrees of freedom, p-value, and, optionally, a mean with 95% confidence interval in square brackets.
}
\description{
Describe t-test results
}
\examples{
t_res<-t.test(rnorm(20, mean = -10, sd=2))
describe.ttest(t_res)
describe.ttest(t_res, show.mean=T)
describe.ttest(t_res, show.mean=T, abs=T)
}
