% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{f.round}
\alias{f.round}
\title{Formatted rounding}
\usage{
f.round(x, digits = 2, strip.lead.zeros = FALSE)
}
\arguments{
\item{x}{A number}

\item{digits}{Number of decimal digits to keep}

\item{strip.lead.zeros}{remove zero before decimal point (default is false)}
}
\value{
Value A number rounded to the specified number of digits
}
\description{
Formatted rounding
}
\examples{
f.round(5.8242)
f.round(5.8251)
f.round(5.82999, digits = 3)
f.round(5.82999, digits = 4)
}
