% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faces-data.R
\docType{data}
\name{faces}
\alias{faces}
\title{Faces identification experiment data}
\format{
A data.table with 2580 observations and 6 variables.
}
\source{
Data accompanying Chetverikov (2015).
}
\usage{
data(faces)
}
\description{
This dataset contains data on response times in a face identification task. The faces were presented for a short time, then two faces (old and new) appeared and participants were asked to choose the old one. The data contains the information about gender (both stimuli gender and participants gender) and answer accuracy and can used to illustrate the mixed models regression approach.
}
\details{
You are free to use it under a Creative Commons Attribution 4.0 International License. When used in research papers, please cite: Chetverikov, A.A. (2015). Linear mixed effects regression in cognitive studies. The Russian Journal of Cognitive Science, 2(1), 41-51. A current copy of the article is available from the Cognitive Research Group publications page: \url{https://cogpsy.ru/publications/}

\itemize{
\item uid. participant ID
\item sid. stimulus ID
\item stim_gender. stimulus gender
\item user_gender. participant's gender
\item correct. response accuracy
\item answerTime. response time
}
}
\references{
Chetverikov, A.A. (2015). Linear mixed effects regression in cognitive studies. The Russian Journal of Cognitive Science, 2(1), 41-51.
}
\keyword{datasets}
