% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{format.results}
\alias{format.results}
\title{Format results}
\usage{
\method{generic}{class}(res_str, type = "pandoc")
}
\arguments{
\item{res_str}{text}

\item{type}{'pandoc' or 'latex'}
}
\value{
\code{res_str} with latex 'emph' tags replaced with pandoc '_'
}
\description{
Internal function used to convert latex-formatted results to pandoc style.
}

