% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{format_results}
\alias{format_results}
\title{Format results}
\usage{
format_results(res_str, type = "pandoc")
}
\arguments{
\item{res_str}{text}

\item{type}{'pandoc', 'latex', or 'plotmath' (the latter is very poorly implemented)}
}
\value{
\code{res_str} with latex 'emph' tags replaced with pandoc '_'
}
\description{
Internal function used to convert latex-formatted results to pandoc style.
}
