% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_grob_element}
\alias{get_grob_element}
\title{Extract grob element by name}
\usage{
get_grob_element(ggplot_obj, el = "guide-box")
}
\arguments{
\item{ggplot_obj}{plot to extract grob from}

\item{el}{element name to extract}
}
\value{
grob element
}
\description{
Extract grob element by name
}
