% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_superb_ci}
\alias{get_superb_ci}
\title{Get within-subject CI using the superb package}
\usage{
get_superb_ci(
  data,
  wid,
  within,
  value_var,
  between = NULL,
  adjustments = list(purpose = "single", decorrelation = "CM"),
  errorbar = "CI",
  drop_NA_subj = FALSE,
  drop_missing_levels = TRUE,
  aggr_fun = mean,
  debug = FALSE,
  ...
)
}
\arguments{
\item{data}{dataframe to use}

\item{wid}{subject variable or another clustering variable (string)}

\item{within}{within-subject variables (vector of strings)}

\item{value_var}{dependent variable (string)}

\item{between}{between-subject variables (vector of strings; default: NULL)}

\item{adjustments}{adjustment settings as used for \link[superb:superbData]{superb::superbData} (default: single CI estimates, Cousineau-Morey adjustment)}

\item{errorbar}{type of error bars to use (CI/SE)}

\item{drop_NA_subj}{should subjects with NA values be dropped? (default: FALSE)}

\item{drop_missing_levels}{should the missing levels of the variables in within/between be dropped? (default: TRUE)}

\item{debug}{output additional debugging info (default: FALSE)}

\item{...}{additional parameters passed to \link{apa_format_mean_conf}}
}
\value{
dataframe with computed CIs
}
\description{
Get within-subject CI using the superb package
}
\examples{
data(faces)
get_superb_ci(faces, "uid", "stim_gender", "answerTime")

}
