% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{Arrange several plots with one legend and/or one x-axis title}
\usage{
grid_arrange_shared_legend(
  ...,
  stack = "v",
  one_sub = F,
  heights = F,
  widths = F,
  one_x_axis = F,
  legend_pos = "b",
  share_legend = T
)
}
\arguments{
\item{...}{- two or more ggplot2 plots}

\item{stack}{- stack direction, "h" (horizontal) or "v" (vertical)}

\item{one_sub}{- one x-axis label for all plots? T/F}

\item{heights}{- a vector of heights for vertically arranged plots}

\item{widths}{- a vector of widths for horizontally arranged plots}

\item{one_x_axis}{- one x-axis for all plots? T/F}

\item{legend}{position - "t" (top),"b" (bottom), "l" (left), or "r" (right)}
}
\value{
plots arranged with grid.arrange
}
\description{
Based on a function from https://github.com/hadley/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs
}
\examples{
load.libs(c('data.table','ggplot2','scales','grid'))
data(faces)

p1<-plot.pointrange(faces[correct==1, ], aes(x=user_gender, color=stim_gender, y=answerTime), wid='uid', within_subj=T)+labs(x="Participant's gender", color="Face Gender", y='Untransformed RT')
p2<-plot.pointrange(faces, aes(x=user_gender, color=stim_gender, y=correct), wid='uid')+labs(x="Participant's gender", color="Face Gender" ,y='Accuracy')
grid_arrange_shared_legend(p1, p2, stack = 'horizontal')
grid_arrange_shared_legend(p1, p2, stack = 'vertical')

#one_x_axis=T has a small bug for now; the labels are not aligned properly
grid_arrange_shared_legend(p1+theme(plot.margin=unit(rep(0,4),'lines'), legend.position='top'),p2+theme(plot.margin=unit(rep(0,4),'lines')), stack='v', one_x_axis = T, one_sub = T,  legend_pos='t')

grid_arrange_shared_legend(p1, p2, stack = 'v', heights=c(0.7,0.2))

}
