% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{ins_lmer}
\alias{ins_lmer}
\title{Describe lmer in-text}
\usage{
ins_lmer(obj, term = NULL, digits = 2, adj.digits = TRUE)
}
\arguments{
\item{obj}{LMER model from \link[lme4:lmer]{lme4::lmer}}

\item{term}{model term to describe (a string with the term name or its sequential number)}

\item{digits}{number of digits for B and SD}

\item{adj.digits}{automatically adjusts digits so that B would not show up as "0.00"}
}
\value{
result
}
\description{
A shortcut for apa.glm(..., dtype=4)
}
\examples{
if (requireNamespace("lme4", quietly = TRUE)) {
  # Fit a mixed-effects model
  fm <- lme4::lmer(Reaction ~ Days + (Days | Subject), lme4::sleepstudy)
  
  # Concise in-text reporting
  ins_lmer(fm, "Days")
  ins_lmer(fm, "(Intercept)", digits = 1)
}
}
