% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lengthu}
\alias{lengthu}
\title{Length of unique values}
\usage{
lengthu(x)
}
\arguments{
\item{x}{a vector}
}
\value{
number of unique values in x
}
\description{
Counts unique values
}
\examples{
x <- c(5, 7, 8, 9, 5, 7)
length(x)
lengthu(x)
}
