% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lmer.fixef}
\alias{lmer.fixef}
\title{Get nice matrix of fixed effects from lmer}
\usage{
lmer.fixef(fit.lmer)
}
\arguments{
\item{fit.lmer}{fitted `lme4::lmer` model object}
}
\value{
result
}
\description{
Get nice matrix of fixed effects from lmer
}
