% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lmer_fixef}
\alias{lmer_fixef}
\title{Get nice matrix of fixed effects from lmer}
\usage{
lmer_fixef(fit.lmer)
}
\arguments{
\item{fit.lmer}{fitted \code{lme4::lmer} model object}
}
\value{
result
}
\description{
Get nice matrix of fixed effects from lmer
}
