% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{lmer_with_julia}
\alias{lmer_with_julia}
\title{Run lmer with Julia}
\usage{
lmer_with_julia(myform, dataset)
}
\arguments{
\item{myform}{}

\item{dataset}{}
}
\value{
result
}
\description{
Run lmer with Julia
}

