% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{load.libs}
\alias{load.libs}
\title{Quietly load libraries}
\usage{
load.libs(libs)
}
\arguments{
\item{libs}{a vector of libraries names}
}
\value{
None
}
\description{
Quietly load libraries
}
\examples{
load.libs(c('ggplot2','apastats','Hmisc'))
}

