% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{load_libs}
\alias{load_libs}
\title{Quietly load libraries}
\usage{
load_libs(libs)
}
\arguments{
\item{libs}{a vector of libraries names}
}
\value{
None
}
\description{
Quietly load libraries
}
\examples{
load_libs(c("ggplot2", "apastats2", "Hmisc"))
}
