% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mean.round}
\alias{mean.round}
\title{Rounded mean}
\usage{
\method{mean}{round}(x, ...)
}
\arguments{
\item{x}{a number}

\item{...}{other arguments passed to \link{f.round}}
}
\value{
Mean rounded to the specified number of digits (string)
}
\description{
Mean rounded to the specified number of digits
}
\examples{
mean.round(c(10, 99))
mean.round(c(10, 99, NA))
mean.round(c(10, 99), 2)
}
