% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mean_round}
\alias{mean_round}
\title{Rounded mean}
\usage{
mean_round(x, ...)
}
\arguments{
\item{x}{a number}

\item{...}{other arguments passed to \link{f_round}}
}
\value{
Mean rounded to the specified number of digits (string)
}
\description{
Mean rounded to the specified number of digits
}
\examples{
mean_round(c(10, 99))
mean_round(c(10, 99, NA))
mean_round(c(10, 99), 2)
}
