% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mymean}
\alias{mymean}
\title{Mean with na.rm=T}
\usage{
mymean(x, ...)
}
\arguments{
\item{x}{a vector of numbers}

\item{...}{}
}
\value{
mean of x with NA removed
}
\description{
Mean with na.rm=T
}
\examples{
x<-c(NA, 10,90)
mean(x)
mymean(x)
}
