% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{mysd}
\alias{mysd}
\title{SD with na.rm=T}
\usage{
mysd(x, ...)
}
\arguments{
\item{x}{a vector of numbers}

\item{...}{}
}
\value{
sd of x with NA removed
}
\description{
SD with na.rm=T
}
\examples{
x<-c(NA, 10,90)
sd(x)
mysd(x)
}

