% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mysum}
\alias{mysum}
\title{Sum with na.rm=T}
\usage{
mysum(x, ...)
}
\arguments{
\item{x}{a vector of numbers}

\item{...}{}
}
\value{
sum of x with NA removed
}
\description{
Sum with na.rm=T
}
\examples{
x<-c(NA, 10,90)
sum(x)
mysum(x)
}

