% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{omit_zeroes}
\alias{omit_zeroes}
\title{Omit leading zero from number}
\usage{
omit_zeroes(x, digits = 2)
}
\arguments{
\item{x}{A number}

\item{digits}{Number of decimal digits to keep}
}
\value{
A number without leading zero
}
\description{
Omit leading zero from number
}
\examples{
omit_zeroes(0.2312)
omit_zeroes(0.2312, digits = 3)
omit_zeroes("000.2312", digits = 1)
}
