% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{paste_and}
\alias{paste_and}
\title{Paste several strings, add 'and' before last}
\usage{
paste_and(x, sep = ", ", suffix = "")
}
\arguments{
\item{x}{vector of strings}

\item{sep}{separator (is not used for only two groups)}

\item{suffix}{suffix to append to each value before the separator}
}
\value{
a string iterating the values in x
}
\description{
Paste several strings, add 'and' before last
}
\examples{
data(iris)
# get mean petal width and SD by group
res <- as.vector(by(iris$Sepal.Width, iris$Species, apa_mean_sd))
res
paste_and(res)
paste_and(res, sep = ";")

data(memory_noise)
# get bias and SD in the two experiment samples
res <- as.vector(by(
  memory_noise$bias_percent,
  memory_noise$experiment,
  apa_mean_sd
))
res
# no comma with two groups
paste_and(res)
paste_and(res, suffix = " percentage points")

}
