% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{round_p}
\alias{round_p}
\title{Round \emph{p}-value}
\usage{
round_p(
  values,
  include.rel = 1,
  digits = 3,
  strip.lead.zeros = TRUE,
  replace.very.small = 0.001
)
}
\arguments{
\item{values}{a vector of p-values}

\item{include.rel}{include relation sign}

\item{digits}{a number of decimal digits}

\item{strip.lead.zeros}{remove zero before decimal point}

\item{replace.very.small}{replace values lower than this criteria (NULL to keep values as is)}
}
\value{
Formatted p-value
}
\description{
If p-value is <= 0.001, returns ".001" else returns p-value rounded to the specified number of digits, optionally including relation sign ("<" or "=").
}
\examples{
p_values <- c(0.025, 0.0001, 0.001, 0.568)
round_p(p_values)
round_p(p_values, digits = 2)
round_p(p_values, include.rel = FALSE)
round_p(p_values, include.rel = FALSE, strip.lead.zeros = FALSE)
round_p(p_values, include.rel = FALSE, strip.lead.zeros = FALSE, replace.very.small = 0.01)

}
