% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{scale_shape_ac}
\alias{scale_shape_ac}
\title{Nice palette of shapes}
\usage{
scale_shape_ac(...)
}
\arguments{
\item{...}{}
}
\value{
scale with nice shapes palette
}
\description{
Nice palette of shapes
}

