% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scale_y_exp}
\alias{scale_y_exp}
\title{Exponentiate scale}
\usage{
scale_y_exp(digits = 0, ...)
}
\arguments{
\item{digits}{number of digits to use in labels}

\item{...}{other arguments passed to scale_y_continuous}
}
\value{
scale
}
\description{
Helps to transform log-scale back to normal
}
