% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sd.nn}
\alias{sd.nn}
\title{SD with na.rm=T}
\usage{
sd.nn(x, ...)
}
\arguments{
\item{x}{a vector of numbers}

\item{...}{other arguments passed to sd}
}
\value{
sd of x with NA removed
}
\description{
SD with na.rm=T
}
\examples{
x <- c(NA, 10, 90)
sd(x)
sd.nn(x)
}
