% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sd.round}
\alias{sd.round}
\title{Rounded SD}
\usage{
sd.round(x, ...)
}
\arguments{
\item{x}{a number}

\item{digits}{decimal places}
}
\value{
Mean rounded to the specified number of digits (string)
}
\description{
SD rounded to the specified number of digits
}
\examples{
sd.round(c(10,99))
sd.round(c(10,99,NA))
sd.round(c(10,99), 2)
}

