% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sd_round}
\alias{sd_round}
\title{Rounded SD}
\usage{
sd_round(x, ...)
}
\arguments{
\item{x}{a number}

\item{...}{other arguments passed to \link{f_round}}
}
\value{
Mean rounded to the specified number of digits (string)
}
\description{
SD rounded to the specified number of digits
}
\examples{
sd_round(c(10, 99))
sd_round(c(10, 99, NA))
sd_round(c(10, 99), 2)
}
