% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{seq_mean}
\alias{seq_mean}
\title{Get the means between the points of a sequence}
\usage{
seq_mean(x)
}
\arguments{
\item{x}{a vector of numeric values}
}
\value{
means between consecutive values
}
\description{
Get the means between the points of a sequence
}
\examples{

seq_mean(c(1, 3, 5, 7, 10))

}
