% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sum_nn}
\alias{sum_nn}
\title{Sum with na.rm=T}
\usage{
sum_nn(x, ...)
}
\arguments{
\item{x}{a vector of numbers}

\item{...}{other arguments passed to sum}
}
\value{
sum of x with NA removed
}
\description{
Sum with na.rm=T
}
\examples{
x <- c(NA, 10, 90)
sum(x)
sum_nn(x)
}
