% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{table.mean.conf}
\alias{table.mean.conf}
\title{Show table of means and confidence interval}
\usage{
table.mean.conf(x, digits = 2, binom = F, ...)
}
\arguments{
\item{x}{}

\item{digits}{}

\item{binom}{}

\item{...}{}
}
\value{
result
}
\description{
Show table of means and confidence interval
}

