% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_functions.R
\name{apa.anova.merMod}
\alias{apa.anova.merMod}
\title{Describe lmerTest anova results}
\usage{
\method{apa}{anova.merMod}(obj, term, f.digits = 2, ...)
}
\arguments{
\item{obj}{lmerTest anova results}

\item{term}{model term to describe (a string with the term name or its sequential number)}

\item{f.digits}{decimal digits for F value}

\item{...}{other parameters passed to \link{format.results}}
}
\value{
formatted string describing the results of anova
}
\description{
Describe lmerTest anova results
}
\examples{
if (requireNamespace("lmerTest", quietly = TRUE)) {
  # Sample data
  data(sleepstudy, package = "lme4")
  
  # Fit a mixed-effects model
  fit <- lmerTest::lmer(Reaction ~ Days + (1 + Days | Subject), sleepstudy)
  
  # ANOVA table
  afit <- anova(fit)
  
  # Format results in APA style
  apa(afit, "Days")
  apa(afit, "Days", f.digits = 3)
}
}
