% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apa.anova}
\alias{apa.anova}
\title{Describe the results of \link[car:Anova]{car::Anova} model or of a \link[stats:anova]{stats::anova} model comparison}
\usage{
\method{apa}{anova}(obj, term = 2, f.digits = 2, ...)
}
\arguments{
\item{obj}{ANOVA results from \link[car:Anova]{car::Anova} or \link[stats:anova]{stats::anova}}

\item{term}{model term to describe (a string with the term name or its sequential number, default: 2)}

\item{f.digits}{number of digits in the results (default: 2)}

\item{...}{other parameters passed to \link{format.results}}
}
\value{
Formatted string with F (or chi2), df, and p
}
\description{
Describe the results of \link[car:Anova]{car::Anova} model or of a \link[stats:anova]{stats::anova} model comparison
}
\details{
When using model comparison version, \code{term} can only be a number.
}
\examples{
# model comparison version
mod1 <- lm(conformity ~ 1, data = carData::Moore)
mod2 <- lm(conformity ~ fcategory, data = carData::Moore, contrasts = list(fcategory = contr.sum))
mod3 <- lm(conformity ~ fcategory * partner.status, data = carData::Moore, contrasts = list(fcategory = contr.sum, partner.status = contr.sum))
anova_obj <- anova(mod1, mod2, mod3)

anova_obj
apa(anova_obj)
apa(anova_obj, 3)

# car::Anova version
mod <- lm(conformity ~ fcategory * partner.status, data = carData::Moore, contrasts = list(fcategory = contr.sum, partner.status = contr.sum))
afit <- car::Anova(mod)

afit
apa(afit, "fcategory")
apa(afit, 2, 4)

}
