% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apa.aov}
\alias{apa.aov}
\title{Describe \link[stats:aov]{stats::aov} results}
\usage{
\method{apa}{aov}(obj, term = term, sstype = 2, ...)
}
\arguments{
\item{obj}{fitted \link[stats:aov]{stats::aov} model}

\item{term}{model term to describe (a string with the term name or its sequential number)}

\item{sstype}{anova SS type (e.g., 2 or 3)}

\item{...}{other parameters passed to apa.anova}
}
\value{
formatted string with F(df_numerator, df_denomiator) = F_value, p =/< p_value
}
\description{
Describe \link[stats:aov]{stats::aov} results
}
\examples{
fit <- aov(answerTime ~ stim_gender*user_gender, data = faces)
apa(fit, 'stim_gender')

}
