% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_functions.R
\name{apa.aov}
\alias{apa.aov}
\title{Describe \link[stats:aov]{stats::aov} results}
\usage{
\method{apa}{aov}(obj, term, sstype = 2, ...)
}
\arguments{
\item{obj}{fitted \link[stats:aov]{stats::aov} model}

\item{term}{model term to describe (a string with the term name or its sequential number)}

\item{sstype}{anova SS type (e.g., 2 or 3)}

\item{...}{other parameters passed to \link{apa.anova}}
}
\value{
formatted string with F(df_numerator, df_denomiator) = F_value, p =/< p_value
}
\description{
Describe \link[stats:aov]{stats::aov} results
}
\examples{
# Using the mtcars dataset
fit <- aov(mpg ~ cyl * am, data = mtcars)
apa(fit, 'cyl')
apa(fit, 'am')
apa(fit, 'cyl:am')

# Using a different SS type
apa(fit, 'cyl', sstype = 3)
}
