% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{apa.BFBayesFactor}
\alias{apa.BFBayesFactor}
\title{Describe BayesFactor results}
\usage{
\method{apa}{BFBayesFactor}(obj, digits = 2, top_limit = 10000, convert_to_power = TRUE, ...)
}
\arguments{
\item{obj}{an object of \link[BayesFactor:BayesFactor-package]{BayesFactor::BayesFactor} class \code{BFBayesFactor}}

\item{digits}{number of digits to use}

\item{top_limit}{numbers above that limit (or below the digits limit) will be converted to exponential notation (if convert_to_power is TRUE)}

\item{convert_to_power}{enable or disable converting of very small or very large numbers to exponential notation}

\item{...}{other parameters passed to \link{format_results}}
}
\value{
string describing the result
}
\description{
Describe BayesFactor results
}
\note{
Code for converting to exponential notation is based on http://dankelley.github.io/r/2015/03/22/scinot.html
}
\examples{
if (requireNamespace("BayesFactor", quietly = TRUE)) {
  # Load the puzzle data from BayesFactor package
  data(puzzles, package = "BayesFactor")
  
  # Run Bayesian ANOVA
  bfs <- BayesFactor::anovaBF(RT ~ shape * color + ID, data = puzzles, progress = FALSE)
  apa(bfs[1])
  apa(bfs[2] / bfs[14])
}
}
