% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{apa.chisq.test}
\alias{apa.chisq.test}
\title{Format chi-square test results in APA style}
\usage{
\method{apa}{chisq.test}(obj, v = TRUE, addN = TRUE, ...)
}
\arguments{
\item{obj}{A result from \link[stats:chisq.test]{stats::chisq.test}}

\item{v}{Add Cramer's V (default: TRUE)}

\item{addN}{Add N (default: TRUE)}

\item{...}{Other parameters passed to \link{format.results}}
}
\value{
Formatted results string
}
\description{
Format chi-square test results in APA style
}
\examples{
# Create a contingency table
M <- as.table(rbind(c(762, 327, 468), c(484, 239, 477)))
dimnames(M) <- list(
  gender = c("F", "M"),
  party = c("Democrat", "Independent", "Republican")
)
chi_result <- chisq.test(M)
apa(chi_result)
}
