% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{apa.dip.test}
\alias{apa.dip.test}
\title{Describe Hartigans' dip test results}
\usage{
\method{apa}{dip.test}(obj, ...)
}
\arguments{
\item{obj}{a result from \link[diptest:dip.test]{diptest::dip.test}}

\item{...}{other parameters passed to \link{format.results}}
}
\value{
formatted results string
}
\description{
Describe Hartigans' dip test results
}
\examples{
if (requireNamespace("diptest", quietly = TRUE)) {
  # Generate bimodal data
  bimodal_data <- c(rnorm(100, -2, 1), rnorm(100, 2, 1))
  
  # Perform dip test
  dip_result <- diptest::dip.test(bimodal_data)
  
  # Format results in APA style
  apa(dip_result)
}
}
