% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apa.ezanova}
\alias{apa.ezanova}
\title{Describe ezANOVA results
Provides formatted string like \emph{F}(DFn, DFd) = ..., \emph{p} ..., eta2 = ... based on ezANOVA results}
\usage{
\method{apa}{ezanova}(
  ezfit,
  term,
  include_eta = TRUE,
  spher_corr = TRUE,
  eta_digits = 2,
  f_digits = 2,
  df_digits = 0,
  append_to_table = FALSE,
  ...
)
}
\arguments{
\item{ezfit}{ezANOVA object}

\item{term}{model term to describe (a string with the term name or its sequential number)}

\item{include_eta}{add eta^2 for the model (default: True)}

\item{spher_corr}{use sphericity corrections    (default: True)}

\item{eta_digits}{number of digits to use for eta^2 (default: 2)}

\item{f_digits}{number of digits to use for F (default: 2)}

\item{df_digits}{number of digits to use for df (default: 0)}

\item{append_to_table}{should the results be added to the original ezANOVA table (default: False)}

\item{...}{other parameters passed to \link{format.results}}
}
\value{
string with formatted results
}
\description{
Describe ezANOVA results
Provides formatted string like \emph{F}(DFn, DFd) = ..., \emph{p} ..., eta2 = ... based on ezANOVA results
}
\examples{

data(faces)
faces$uid <- factor(faces$uid)
ez_res <- ez::ezANOVA(faces[faces$correct == 1], dv = answerTime, 
                      wid = uid, within = stim_gender, between = user_gender)
apa(ez_res, "user_gender")
apa(ez_res, "user_gender", eta_digits = 3)
apa(ez_res, "user_gender:stim_gender", eta_digits = 3)
apa(ez_res, 3, eta_digits = 3)
}
