% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_functions.R
\name{apa.ezANOVA}
\alias{apa.ezANOVA}
\title{Describe ezANOVA results}
\usage{
\method{apa}{ezANOVA}(
  obj,
  term,
  include_eta = TRUE,
  spher_corr = TRUE,
  eta_digits = 2,
  f_digits = 2,
  df_digits = 0,
  append_to_table = FALSE,
  ...
)
}
\arguments{
\item{obj}{ezANOVA object from \link[ez:ezANOVA]{ez::ezANOVA}}

\item{term}{model term to describe (a string with the term name or its sequential number)}

\item{include_eta}{add eta^2 for the model (default: TRUE)}

\item{spher_corr}{use sphericity corrections    (default: TRUE)}

\item{eta_digits}{number of digits to use for eta^2 (default: 2)}

\item{f_digits}{number of digits to use for F (default: 2)}

\item{df_digits}{number of digits to use for df (default: 0)}

\item{append_to_table}{should the results be added to the original ezANOVA table (default: FALSE)}

\item{...}{other parameters passed to \link{format.results}}
}
\value{
string with formatted results
}
\description{
Provides formatted string like \emph{F}(DFn, DFd) = ..., \emph{p} ..., eta2 = ... based on ezANOVA results
}
\examples{
if (requireNamespace("ez", quietly = TRUE) && requireNamespace("afex", quietly = TRUE)) {
  # Using the Stroop dataset from afex package
  data(stroop, package = "afex")
  
  # Run ezANOVA
  ez_res <- ez::ezANOVA(
    data = stroop[!is.na(stroop$rt),],
    dv = rt,
    wid = pno,
    within = condition,
    detailed = TRUE
  )
  
  # Format results in APA style
  apa(ez_res, "condition")
  apa(ez_res, "condition", eta_digits = 3)
  apa(ez_res, "condition", include_eta = FALSE)
}
}
