% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apa.glm}
\alias{apa.glm}
\title{Describe regression model (GLM, GLMer, lm, lm.circular, ...)}
\usage{
\method{apa}{glm}(
  fit,
  term = NULL,
  dtype = 1,
  b.digits = 2,
  t.digits = 2,
  test.df = FALSE,
  p.as.number = FALSE,
  term.pattern = NULL,
  eff.size = FALSE,
  adj.digits = FALSE,
  ...
)
}
\arguments{
\item{fit}{model object}

\item{term}{model term to describe (a string with the term name or its sequential number)}

\item{dtype}{description type (1: t, p;  2: B(SE), p; 3: B, SE, t, p; or other: B (SE), t)}

\item{b.digits}{how many digits to use for \emph{B} and \emph{SE}}

\item{t.digits}{how many digits to use for \emph{t}}

\item{test.df}{should we include degrees of freedom in description?}

\item{p.as.number}{should the p-values be transformed to numbers (T) or shown as strings (F)?}

\item{term.pattern}{return only the model terms matching the regex pattern (grepl is used)}

\item{eff.size}{should we include effect size (currently implemented only for simple regression)?}

\item{adj.digits}{automatically adjusts digits so that B or SE would not show up as "0.00"}

\item{...}{other parameters passed to \link{format.results}}
}
\value{
result
}
\description{
Describe regression model (GLM, GLMer, lm, lm.circular, ...)
}
\examples{

Animals <- MASS::Animals
Animals$body <- log(Animals$body)
Animals$brain <- log(Animals$brain)
fit <- lm(brain ~ body, Animals)
apa(fit, "body")
apa(fit, "(Intercept)")
apa(fit, "body", 2)
apa(fit, "body", 3)
apa(fit, "body", 4)
apa(fit, "body", 3, test.df = 1)
apa(fit)
\dontrun{
require(rockchalk)
apa(fit, "body", 4, eff.size = TRUE)
}
}
