% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apa.htest}
\alias{apa.htest}
\alias{apa.chisq}
\alias{apa.dip.test}
\title{Describe the results of $\chi^2$ or Hartigans' dip tests}
\usage{
\method{apa}{htest}(obj, ...)

\method{apa}{chisq}(obj, v = TRUE, addN = TRUE, ...)

\method{apa}{dip.test}(obj, ...)
}
\arguments{
\item{obj}{a result from \link[stats:chisq.test]{stats::chisq.test} or \link[diptest:dip.test]{diptest::dip.test}}

\item{...}{other parameters passed to \link{format.results}}

\item{v}{add Cramer's V (default: T)}

\item{addN}{add N (default: T)}
}
\value{
formatted results string
}
\description{
Describe the results of $\chi^2$ or Hartigans' dip tests
}
\section{Functions}{
\itemize{
\item \code{apa(chisq)}: Describe Pearson $\\chi^2$ results

\item \code{apa(dip.test)}: Describe Hartigans' dip test results

}}
\examples{

## From chisq.test help page
## ---------------------
## From Agresti(2007) p.39
M <- as.table(rbind(c(762, 327, 468), c(484, 239, 477)))
dimnames(M) <- list(
  gender = c("FALSE", "M"),
  party = c("Democrat", "Independent", "Republican")
)
(Xsq <- chisq.test(M)) # Prints test summary
## ----------------------
apa(Xsq)


requireNamespace('diptest')
data(statfaculty, package = 'diptest')
(d.t <- dip.test(statfaculty))
apa(d.t)

}
