% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{apa.list}
\alias{apa.list}
\title{Format results from list objects in APA style}
\usage{
\method{apa}{list}(obj, ...)
}
\arguments{
\item{obj}{A list object containing statistical results}

\item{...}{Additional arguments passed to the appropriate method}
}
\value{
A formatted string with statistical results in APA style
}
\description{
This function detects the type of statistical results stored in a list object
and redirects to the appropriate APA formatting method. Currently supports
\link{ezANOVA} results (\link{apa.ezANOVA}).
}
