% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{apa.r}
\alias{apa.r}
\title{Format correlation test results in APA style}
\usage{
\method{apa}{r}(obj, ...)
}
\arguments{
\item{obj}{An object from \link[stats:cor.test]{stats::cor.test}}

\item{...}{Other arguments passed to \link{format.results}}
}
\value{
A string with correlation coefficient, degrees of freedom, and p-value
}
\description{
Format correlation test results in APA style
}
\examples{
# Pearson correlation
x <- rnorm(40)
y <- x * 0.6 + rnorm(40, 0, 0.8)
rc <- cor.test(x, y)
apa.r(rc)
}
