% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apa}
\alias{apa}
\title{Describe an object in APA style}
\usage{
apa(obj, ...)
}
\arguments{
\item{obj}{an object to describe}

\item{...}{other arguments passed to}
}
\value{
description in APA (American Psychological Association) format
}
\description{
Describe an object in APA style
}
