% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{apa.roc.test}
\alias{apa.roc.test}
\alias{apa.roc.diff}
\title{Describe differences between ROC curves}
\usage{
\method{apa}{roc.test}(obj)

\method{apa}{roc.diff}(obj)
}
\arguments{
\item{obj}{a difference between the ROC curves from \link[pROC:roc.test]{pROC::roc.test}}
}
\value{
result
}
\description{
Describe differences between ROC curves
}
\examples{
if (requireNamespace("pROC", quietly = TRUE)) {
  # Create sample data
  set.seed(42)
  n <- 100
  group <- factor(sample(c(0, 1), n, replace = TRUE))
  test1 <- rnorm(n, mean = 1 * (as.numeric(group) - 1))
  test2 <- rnorm(n, mean = 0.7 * (as.numeric(group) - 1))
  
  # Create ROC curves
  roc1 <- pROC::roc(group, test1)
  roc2 <- pROC::roc(group, test2)
  
  # Test difference between ROC curves
  roc_diff <- pROC::roc.test(roc1, roc2)
  
  # Format results in APA style
  apa(roc_diff)
}
}
