% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{apa_mean_and_t}
\alias{apa_mean_and_t}
\title{Compare means with t-tests and format in APA style}
\usage{
apa_mean_and_t(
  x,
  by,
  which.mean = 1,
  digits = 2,
  paired = FALSE,
  eff.size = FALSE,
  abs = FALSE,
  aggregate_by = NULL,
  transform.means = NULL,
  ...
)
}
\arguments{
\item{x}{Dependent variable (numeric vector)}

\item{by}{Independent variable (factor with 2 levels)}

\item{which.mean}{Which means to show (0=none, 1=first, 2=second, 3=both)}

\item{digits}{Number of digits in results (default: 2)}

\item{paired}{Should it be a paired test (default: FALSE)}

\item{eff.size}{Should we include effect size (default: FALSE)}

\item{abs}{Should we show the absolute value if the t-test (T) or keep its sign (FALSE, default)}

\item{aggregate_by}{Do the aggregation by the third variable(s): either NULL (default), a single vector variable, or a list of variables to aggregate by.}

\item{transform.means}{A function to transform means and confidence intervals (default: NULL)}

\item{...}{Additional arguments passed to \link{format.results}}
}
\value{
A formatted string with t-test results in APA style
}
\description{
Compare means with t-tests and format in APA style
}
\examples{
# Using simulated data
rt <- rnorm(100)
gr <- factor(rep(c("A", "B"), each = 50))
apa_mean_and_t(rt, gr)
apa_mean_and_t(rt, gr, which.mean = 3)
apa_mean_and_t(rt, gr, eff.size = TRUE)
}
