% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{apastats-deprecated}
\alias{apastats-deprecated}
\alias{describe.ttest}
\alias{describe.r}
\alias{describe.chi}
\alias{describe.Anova}
\alias{describe.aov}
\alias{describe.anova}
\alias{describe.bf}
\alias{describe.brm}
\alias{describe.dip.test}
\alias{describe.emmeans}
\alias{describe.ezanova}
\alias{describe.ezstats}
\alias{describe.glm}
\alias{describe.lht}
\alias{describe.lmer}
\alias{describe.lmert}
\alias{describe.lmtaov}
\alias{describe.lsmeans}
\alias{describe.roc.diff}
\alias{describe.mean.sd}
\alias{describe.mean.conf}
\alias{describe.binom.mean.conf}
\alias{describe.mean.and.t}
\alias{mymean}
\alias{mysum}
\alias{mysd}
\alias{mean.nn}
\alias{sd.nn}
\alias{sum.nn}
\alias{drop.empty.cols}
\alias{binom.ci}
\alias{mean.round}
\alias{sd.round}
\alias{load.libs}
\alias{lmer.fixef}
\alias{omit.zeroes}
\alias{f.round}
\alias{base.breaks}
\alias{base.breaks.x}
\alias{base.breaks.y}
\alias{plot.pointrange}
\title{Deprecated functions in the apastats package}
\usage{
describe.ttest(t, ...)

describe.r(rc, ...)

describe.chi(tbl, ...)

describe.Anova(afit, term, f.digits = 2, ...)

describe.aov(fit, term, sstype = 2, ...)

describe.anova(anova_res, rown = 2, f.digits = 2, ...)

describe.bf(bf, digits = 2, top_limit = 10000, convert_to_power = TRUE, ...)

describe.brm(
  mod,
  term,
  trans = NULL,
  digits = 2,
  eff.size = FALSE,
  eff.size.type = "r",
  nsamples = 100,
  ci.type = "HPDI",
  ...
)

describe.dip.test(x, ...)

describe.emmeans(obj, term, dtype = "B", df = FALSE, ...)

describe.ezanova(
  ezfit,
  term,
  include_eta = TRUE,
  spher_corr = TRUE,
  eta_digits = 2,
  f_digits = 2,
  df_digits = 0,
  append_to_table = FALSE,
  ...
)

describe.ezstats(ezstats_res, term = 1, ...)

describe.glm(
  fit,
  term = NULL,
  dtype = 1,
  b.digits = 2,
  t.digits = 2,
  test.df = FALSE,
  p.as.number = FALSE,
  term.pattern = NULL,
  eff.size = FALSE,
  adj.digits = FALSE,
  ...
)

describe.lht(hyp, ...)

describe.lmer(
  fm,
  pv,
  digits = c(2, 2, 2),
  incl.rel = 0,
  dtype = "B",
  incl.p = TRUE
)

describe.lmert(sfit, factor, dtype = "t", ...)

describe.lmtaov(afit, term, f.digits = 2, ...)

describe.lsmeans(obj, term, dtype = "B", df = FALSE, ...)

describe.roc.diff(roc_diff)

describe.mean.sd(
  x = NULL,
  m = NULL,
  sd = NULL,
  digits = 2,
  dtype = "p",
  m_units = "",
  sd_units = "",
  ...
)

describe.mean.conf(
  x,
  bootCI = TRUE,
  addCI = FALSE,
  digits = 2,
  transform.means = NULL,
  ...
)

describe.binom.mean.conf(x, digits = 2, ...)

describe.mean.and.t(
  x,
  by,
  which.mean = 1,
  digits = 2,
  paired = FALSE,
  eff.size = FALSE,
  abs = FALSE,
  aggregate_by = NULL,
  transform.means = NULL,
  ...
)

mymean(...)

mysum(...)

mysd(...)

\method{mean}{nn}(x, ...)

sd.nn(x, ...)

\method{sum}{nn}(x, ...)

drop.empty.cols(df)

binom.ci(x)

\method{mean}{round}(x, ...)

sd.round(x, ...)

load.libs(libs)

lmer.fixef(fit.lmer)

omit.zeroes(x, digits = 2)

f.round(x, digits = 2, strip.lead.zeros = FALSE)

base.breaks(x, scale = "x", addSegment = TRUE, ...)

base.breaks.x(x, addSegment = TRUE, ...)

base.breaks.y(x, addSegment = TRUE, ...)

\method{plot}{pointrange}(
  data,
  mapping,
  pos = position_dodge(0.3),
  pointsize = I(3),
  linesize = I(1),
  pointfill = I("white"),
  pointshape = NULL,
  within_subj = FALSE,
  wid = "uid",
  bars = "ci",
  withinvars = NULL,
  betweenvars = NULL,
  x_as_numeric = FALSE,
  custom_geom_before = NULL,
  connecting_line = FALSE,
  pretty_breaks_y = FALSE,
  pretty_y_axis = FALSE,
  exp_y = FALSE,
  print_aggregated_data = FALSE,
  do_aggregate = FALSE,
  add_margin = FALSE,
  margin_label = "all",
  margin_x_vals = NULL,
  bars_instead_of_points = FALSE,
  geom_bar_params = list(),
  add_jitter = FALSE,
  individual_points_params = list(),
  drop_NA_subj = FALSE,
  design = "between",
  debug = FALSE
)
}
\description{
These functions are deprecated and will be removed in a future version.
Please use their replacements as indicated.

\itemize{
\item \code{describe.ttest} -> \link{apa} (for t.test objects)
\item \code{describe.r} -> \link{apa} (for cor.test objects)
\item \code{describe.chi} -> \link{apa} (for chisq.test objects)
\item \code{describe.mean.sd} -> \link{apa_mean_sd}
\item \code{describe.mean.conf} -> \link{apa_mean_conf}
\item \code{describe.binom.mean.conf} -> \link{apa_binom_mean_conf}
\item \code{describe.mean.and.t} -> \link{apa_mean_and_t}
\item \code{describe.Anova} -> \link{apa} (for Anova objects)
\item \code{describe.aov} -> \link{apa} (for aov objects)
\item \code{describe.anova} -> \link{apa} (for anova objects)
\item \code{describe.bf} -> \link{apa} (for BFBayesFactor objects)
\item \code{describe.brm} -> \link{apa} (for brmsfit objects)
\item \code{describe.dip.test} -> \link{apa} (for dip.test objects)
\item \code{describe.emmeans} -> \link{apa} (for emmeans objects)
\item \code{describe.ezanova} -> \link{apa} (for ezanova objects)
\item \code{describe.ezstats} -> \link{apa} (for ezstats objects)
\item \code{describe.glm} -> \link{apa} (for glm/lm objects)
\item \code{describe.lht} -> \link{apa} (for linearHypothesis objects)
\item \code{describe.lmer} -> \link{apa} (for lmer objects)
\item \code{describe.lmert} -> \link{apa} (for lmerTest objects)
\item \code{describe.lmtaov} -> \link{apa} (for lmerTest anova objects)
\item \code{describe.lsmeans} -> \link{apa} (for lsmeans objects)
\item \code{describe.roc.diff} -> \link{apa} (for ROC difference objects)
\item \code{mean.nn} -> \link{mean_nn}
\item \code{sd.nn} -> \link{sd_nn}
\item \code{sum.nn} -> \link{sum_nn}
\item \code{drop.empty.cols} -> \link{drop_empty_cols}
\item \code{binom.ci} -> \link{binom_ci}
\item \code{mean.round} -> \link{mean_round}
\item \code{sd.round} -> \link{sd_round}
\item \code{load.libs} -> \link{load_libs}
\item \code{lmer.fixef} -> \link{lmer_fixef}
\item \code{omit.zeroes} -> \link{omit_zeroes}
\item \code{f.round} -> \link{f_round}
\item \code{base.breaks} -> \link{base_breaks}
\item \code{base.breaks.x} -> \link{base_breaks_x}
\item \code{base.breaks.y} -> \link{base_breaks_y}
\item \code{plot.pointrange} -> \link{plot_pointrange}
\item \code{mymean} -> \link{mean_nn}
\item \code{mysum} -> \link{sum_nn}
\item \code{mysd} -> \link{sd_nn}
}
}
\keyword{internal}
