% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{package}
\name{apastats-deprecated}
\alias{apastats-deprecated}
\alias{mymean}
\alias{mysum}
\alias{mysd}
\title{Deprecated function(s) in the apastats package}
\usage{
mymean(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older version of
the apastats package. They will removed in later releases.
}
\section{Details}{

\tabular{rl}{
  \code{mymean} \tab now a synonym for \code{\link{mean.nn}}\cr
  \code{mysum} \tab now a synonym for \code{\link{sum.nn}}\cr
  \code{mysd} \tab now a synonym for \code{\link{sd.nn}}\cr
}
}

