% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{apastats-deprecated}
\alias{apastats-deprecated}
\alias{describe.ttest}
\alias{describe.r}
\alias{describe.chi}
\alias{describe.Anova}
\alias{describe.aov}
\alias{describe.anova}
\alias{describe.bf}
\alias{describe.brm}
\alias{describe.dip.test}
\alias{describe.emmeans}
\alias{describe.ezanova}
\alias{describe.ezstats}
\alias{describe.glm}
\alias{describe.lht}
\alias{describe.lmer}
\alias{describe.lmert}
\alias{describe.lmtaov}
\alias{describe.lsmeans}
\alias{describe.roc.diff}
\alias{describe.mean.sd}
\alias{describe.mean.conf}
\alias{describe.binom.mean.conf}
\alias{describe.mean.and.t}
\alias{mymean}
\alias{mysum}
\alias{mysd}
\title{Deprecated functions in the apastats package}
\usage{
describe.ttest(t, ...)

describe.r(rc, ...)

describe.chi(tbl, ...)

describe.Anova(afit, term, f.digits = 2, ...)

describe.aov(fit, term, sstype = 2, ...)

describe.anova(anova_res, rown = 2, f.digits = 2, ...)

describe.bf(bf, digits = 2, top_limit = 10000, convert_to_power = TRUE, ...)

describe.brm(
  mod,
  term,
  trans = NULL,
  digits = 2,
  eff.size = FALSE,
  eff.size.type = "r",
  nsamples = 100,
  ci.type = "HPDI",
  ...
)

describe.dip.test(x, ...)

describe.emmeans(obj, term, dtype = "B", df = FALSE, ...)

describe.ezanova(
  ezfit,
  term,
  include_eta = TRUE,
  spher_corr = TRUE,
  eta_digits = 2,
  f_digits = 2,
  df_digits = 0,
  append_to_table = FALSE,
  ...
)

describe.ezstats(ezstats_res, term = 1, ...)

describe.glm(
  fit,
  term = NULL,
  dtype = 1,
  b.digits = 2,
  t.digits = 2,
  test.df = FALSE,
  p.as.number = FALSE,
  term.pattern = NULL,
  eff.size = FALSE,
  adj.digits = FALSE,
  ...
)

describe.lht(hyp, ...)

describe.lmer(
  fm,
  pv,
  digits = c(2, 2, 2),
  incl.rel = 0,
  dtype = "B",
  incl.p = TRUE
)

describe.lmert(sfit, factor, dtype = "t", ...)

describe.lmtaov(afit, term, f.digits = 2, ...)

describe.lsmeans(obj, term, dtype = "B", df = FALSE, ...)

describe.roc.diff(roc_diff)

describe.mean.sd(
  x = NULL,
  m = NULL,
  sd = NULL,
  digits = 2,
  dtype = "p",
  m_units = "",
  sd_units = "",
  ...
)

describe.mean.conf(
  x,
  bootCI = TRUE,
  addCI = FALSE,
  digits = 2,
  transform.means = NULL,
  ...
)

describe.binom.mean.conf(x, digits = 2, ...)

describe.mean.and.t(
  x,
  by,
  which.mean = 1,
  digits = 2,
  paired = FALSE,
  eff.size = FALSE,
  abs = FALSE,
  aggregate_by = NULL,
  transform.means = NULL,
  ...
)

mymean(...)

mysum(...)

mysd(...)
}
\description{
These functions are deprecated and will be removed in a future version.
Please use their replacements as indicated.

\itemize{
\item \code{describe.ttest} -> \link{apa} (for t.test objects)
\item \code{describe.r} -> \link{apa} (for cor.test objects)
\item \code{describe.chi} -> \link{apa} (for chisq.test objects)
\item \code{describe.mean.sd} -> \link{apa_mean_sd}
\item \code{describe.mean.conf} -> \link{apa_mean_conf}
\item \code{describe.binom.mean.conf} -> \link{apa_binom_mean_conf}
\item \code{describe.mean.and.t} -> \link{apa_mean_and_t}
\item \code{describe.Anova} -> \link{apa} (for Anova objects)
\item \code{describe.aov} -> \link{apa} (for aov objects)
\item \code{describe.anova} -> \link{apa} (for anova objects)
\item \code{describe.bf} -> \link{apa} (for BFBayesFactor objects)
\item \code{describe.brm} -> \link{apa} (for brmsfit objects)
\item \code{describe.dip.test} -> \link{apa} (for dip.test objects)
\item \code{describe.emmeans} -> \link{apa} (for emmeans objects)
\item \code{describe.ezanova} -> \link{apa} (for ezanova objects)
\item \code{describe.ezstats} -> \link{apa} (for ezstats objects)
\item \code{describe.glm} -> \link{apa} (for glm/lm objects)
\item \code{describe.lht} -> \link{apa} (for linearHypothesis objects)
\item \code{describe.lmer} -> \link{apa} (for lmer objects)
\item \code{describe.lmert} -> \link{apa} (for lmerTest objects)
\item \code{describe.lmtaov} -> \link{apa} (for lmerTest anova objects)
\item \code{describe.lsmeans} -> \link{apa} (for lsmeans objects)
\item \code{describe.roc.diff} -> \link{apa} (for ROC difference objects)
\item \code{mymean} -> \link{mean.nn}
\item \code{mysum} -> \link{sum.nn}
\item \code{mysd} -> \link{sd.nn}
}
}
\keyword{internal}
