% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{base.breaks}
\alias{base.breaks}
\title{Tufte-like breaks for axes}
\usage{
base.breaks(x, scale = "x", addSegment = T, ...)
}
\arguments{
\item{x}{vector of numbers to create breaks from}

\item{scale}{scale (x or y) to create breaks for}

\item{addSegment}{should we add a line to the scale? (T/F)}

\item{...}{other parameters passed to scale_(x or y)_continuous}
}
\value{
scale_(x or y)_continuous with pretty breaks and accompaniying geom_segment if addSegment == T
}
\description{
Tufte-like breaks for axes
}
\examples{

}

