% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.Anova}
\alias{describe.Anova}
\title{Describe Anova results}
\usage{
describe.Anova(afit, term, f.digits = 2, ...)
}
\arguments{
\item{afit}{}

\item{term}{}

\item{f.digits}{}

\item{...}{}
}
\value{
result
}
\description{
Describe Anova results
}

