% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.aov}
\alias{describe.aov}
\title{Describe aov results}
\usage{
describe.aov(fit, term, sstype = 2, ...)
}
\arguments{
\item{fit}{fitted aov model}

\item{term}{model term to describe (a string with the term name or its sequential number)}

\item{sstype}{anova SS type (e.g., 2 or 3)}

\item{...}{other parameters passed to describe.Anova}
}
\value{
formatted string with FALSE(df_numerator, df_denomiator) = F_value, p =/< p_value
}
\description{
Describe aov results
}
