% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.bf}
\alias{describe.bf}
\title{Describe BayesFactor results}
\usage{
describe.bf(bf, digits = 2, top_limit = 10000, convert_to_power = TRUE, ...)
}
\arguments{
\item{bf}{an object of \link[BayesFactor]{BFBayesFactor} class}

\item{digits}{number of digits to use}

\item{top_limit}{numbers above that limit (or below the digits limit) will be converted to exponential notation (if convert_to_power is TRUE)}

\item{convert_to_power}{enable or distable converting of very small or very large numbers to exponential notation}

\item{...}{other parameters passed to \link{format.results}}
}
\value{
string describing the result
}
\description{
Describe BayesFactor results
}
\note{
Code for converting to exponential notation is based on http://dankelley.github.io/r/2015/03/22/scinot.html
}
\examples{

require('BayesFactor')
data(puzzles)

bfs <- anovaBF(RT ~ shape*color + ID, data = puzzles, progress=FALSE)
describe.bf(bfs[1])
describe.bf(bfs[2]/bfs[14])

}
