% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.bimod.test}
\alias{describe.bimod.test}
\title{Describe bimodality test results}
\usage{
describe.bimod.test(x, start_vec = NA, ...)
}
\arguments{
\item{x}{value vector for \link[bimodalitytest]{bimodality.test}}

\item{start_vec}{start vector for \link[bimodalitytest]{bimodality.test}}

\item{...}{other parameters passed to \link{format.results}}
}
\value{
a string with the LR (likelihood ratio) and p
}
\description{
Describe bimodality test results
}
