% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.chi}
\alias{describe.chi}
\title{Describe $chi^2$ results}
\usage{
describe.chi(tbl, v = T, addN = T, ...)
}
\arguments{
\item{tbl}{}

\item{v}{}

\item{addN}{}

\item{...}{}
}
\value{
result
}
\description{
Describe $chi^2$ results
}

