% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.dip.test}
\alias{describe.dip.test}
\title{Describe Hartigans' dip test results}
\usage{
describe.dip.test(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{
result
}
\description{
Describe Hartigans' dip test results
}

