% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.ezanova}
\alias{describe.ezanova}
\title{Describe ezANOVA results
Provides formatted string like _F_(DFn, DFd) = ..., _p_ ..., eta2 = ... based on ezANOVA results}
\usage{
describe.ezanova(
  ezfit,
  term,
  include_eta = T,
  spher_corr = T,
  eta_digits = 2,
  f_digits = 2,
  df_digits = 0,
  ...
)
}
\arguments{
\item{ezfit}{ezANOVA object}

\item{term}{model term to describe (a string with the term name or its sequential number)}

\item{include_eta}{add eta^2 for the model (default: True)}

\item{spher_corr}{use sphericity corrections    (default: True)}

\item{eta_digits}{number of digits to use for eta^2 (default: 2)}

\item{f_digits}{number of digits to use for F (default: 2)}

\item{df_digits}{number of digits to use for df (default: 0)}

\item{...}{other parameters passed to \link{format.results}}
}
\value{
string with formatted results
}
\description{
Describe ezANOVA results
Provides formatted string like _F_(DFn, DFd) = ..., _p_ ..., eta2 = ... based on ezANOVA results
}
\examples{
library(ez)
data(faces)
ez_res <- ezANOVA(faces[faces$correct==1], dv = answerTime, wid = uid, within = stim_gender, between=user_gender)
describe.ezanova(ez_res, 'user_gender')
describe.ezanova(ez_res, 'user_gender', eta_digits = 3)
describe.ezanova(ez_res, 'user_gender:stim_gender', eta_digits = 3)
describe.ezanova(ez_res, 3, eta_digits = 3)
}
