% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.glm}
\alias{describe.glm}
\title{Describe regression model (GLM, GLMer, lm, lm.circular, ...)}
\usage{
describe.glm(fit, term = NULL, dtype = 1, b.digits = 2, t.digits = 2,
  test.df = F, p.as.number = F, term.pattern = NULL, eff.size = F,
  adj.digits = F, ...)
}
\arguments{
\item{fit}{model object}

\item{term}{model term to describe}

\item{dtype}{description type (1: t, p;  2: B(SE), p; 3: B, SE, t, p; or other: B (SE), t)}

\item{b.digits}{how many digits to use for _B_ and _SE_}

\item{t.digits}{how many digits to use for _t_}

\item{test.df}{should we include degrees of freedom in description?}

\item{eff.size}{should we include effect size (currently implemented only for simple regression)?}

\item{adj.digits}{automatically adjusts digits so that B or SE would not show up as "0.00"}

\item{...}{}
}
\value{
result
}
\description{
Describe regression model (GLM, GLMer, lm, lm.circular, ...)
}
\examples{

utils::data(Animals, package = "MASS")
Animals$body<-log(Animals$body)
Animals$brain<-log(Animals$brain)
fit <- lm(brain~body, Animals)
describe.glm(fit, 'body')
describe.glm(fit, '(Intercept)')
describe.glm(fit, 'body', 2)
describe.glm(fit, 'body', 3)
describe.glm(fit, 'body', 4)
describe.glm(fit, 'body', 3, test.df = 1)
describe.glm(fit)
## Not run:
require(rockchalk)
describe.glm(fit, 'body', 4, eff.size=T)
## End(Not run)
}
