% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.lmer}
\alias{describe.lmer}
\title{Describe lmer results}
\usage{
describe.lmer(fm, pv, digits = c(2, 2, 2), incl.rel = 0, dtype = "B",
  incl.p = T)
}
\arguments{
\item{incl.p}{}
}
\value{
result
}
\description{
Describe lmer results
}

