% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.lmer}
\alias{describe.lmer}
\title{Describe lmer results}
\usage{
describe.lmer(
  fm,
  pv,
  digits = c(2, 2, 2),
  incl.rel = 0,
  dtype = "B",
  incl.p = T
)
}
\arguments{
\item{fm}{LMER model from lme4}

\item{pv}{p values table (from Anova)}

\item{digits}{number of digits in results (default: 2)}

\item{incl.rel}{include the relation sign for _p_}

\item{dtype}{description type (B or t)}

\item{incl.p}{include p values}
}
\value{
result
}
\description{
This function is deprecated in favor of \link{describe.glm}
}
