% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.lmert}
\alias{describe.lmert}
\title{Describe lmerTest results}
\usage{
describe.lmert(sfit, factor, dtype = "t", ...)
}
\arguments{
\item{...}{}
}
\value{
result
}
\description{
Describe lmerTest results
}

