% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.lmert}
\alias{describe.lmert}
\title{Describe lmerTest results}
\usage{
describe.lmert(sfit, factor, dtype = "t", ...)
}
\arguments{
\item{sfit}{*summary* object from lmerTest::lmer model}

\item{factor}{name or number of the factor that needs to be describe}

\item{dtype}{description type ("B"/"t")}

\item{...}{other parameters passed to \link{format.results}}
}
\value{
Formatted string
}
\description{
Note that this function uses *summary* object from lmerTest::lmer model and not the model itself (see the example). Otherwise p-values will be computed during the call and everything will be very slow.
}
\examples{

library(lmerTest)
fm <- lmer(Reaction ~ Days + (Days|Subject), sleepstudy)
fms <- summary(fm)
describe.lmert(fms, 'Days')
describe.lmert(fms, '(Intercept)')
describe.lmert(fms, 2)
describe.lmert(fms, 'Days', 'B')

}
