% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.mean.and.t}
\alias{describe.mean.and.t}
\title{Describe two-sample t-test with means and effect sizes}
\usage{
describe.mean.and.t(x, by, which.mean = 1, digits = 2, paired = F,
  eff.size = F, abs = F, ...)
}
\arguments{
\item{x}{dependent variable}

\item{by}{independent variable}

\item{which.mean}{which mean to show (0 - none, 1 - first group (default), 2 - second group, 3 - both)}

\item{digits}{number of digits in results (default: 2)}

\item{paired}{should it be a paired test (default: F)}

\item{eff.size}{should we include effect size (default: F)}

\item{abs}{should we show the absolute value if the t-test (T) or keep its sign (F, default)}

\item{...}{}
}
\value{
result
}
\description{
Describe two-sample t-test with means and effect sizes
}

