% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.mean.conf}
\alias{describe.mean.conf}
\title{Describe mean and confidence intervals}
\usage{
describe.mean.conf(x, digits = 2, ...)
}
\arguments{
\item{x}{}

\item{digits}{}

\item{...}{}
}
\value{
result
}
\description{
Describe mean and confidence intervals
}

