% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{describe.mean.conf}
\alias{describe.mean.conf}
\title{Describe mean and confidence intervals}
\usage{
describe.mean.conf(
  x,
  bootCI = TRUE,
  addCI = FALSE,
  digits = 2,
  transform.means = NULL,
  ...
)
}
\arguments{
\item{x}{value that should be described}

\item{bootCI}{use bootstrapped (T, default) or Gaussian (F) confidence intervals}

\item{addCI}{add "95\% CI =" before confidence intervals (default: F)}

\item{digits}{number of digits to use}

\item{transform.means}{an optional function to transform the means and CI to another scale}

\item{...}{other arguments passed to \link{format.results}}
}
\value{
a string with a mean followed by confidence intervals in square brackets.
}
\description{
Describe mean and confidence intervals
}
\examples{

x <- runif(100, 0, 50)
describe.mean.conf(x)
describe.mean.conf(x, bootCI = FALSE)
describe.mean.conf(x, digits = 5)
describe.mean.conf(x, transform.means = function(val) val * 2)
}
