% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{drop_empty_cols}
\alias{drop_empty_cols}
\title{Drop empty columns from a data.frame}
\usage{
drop_empty_cols(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
data.frame without empty columns
}
\description{
Drop empty (consisting of NA only) columns from a data.frame. Based on https://stackoverflow.com/a/2644009/1344028
}
\examples{
df <- data.frame(x = rnorm(20), y = rep("A", 20), z = rep(NA, 20))
str(df)
df <- drop_empty_cols(df)
str(df)

}
