% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_grob_element}
\alias{get_grob_element}
\title{Extract grob element by name
Not mine, found somewhere}
\usage{
get_grob_element(myggplot, el = "guide-box")
}
\arguments{
\item{el}{}
}
\value{
grob element
}
\description{
Extract grob element by name
Not mine, found somewhere
}

