% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_superb_ci}
\alias{get_superb_ci}
\title{Get within-subject CI using the superb package}
\usage{
get_superb_ci(
  data,
  wid,
  within,
  value_var,
  between = NULL,
  adjustments = list(purpose = "single", decorrelation = "CM"),
  errorbar = "CI"
)
}
\arguments{
\item{data}{dataframe to use}

\item{wid}{subject variable or another clustering variable (string)}

\item{within}{within-subject variables (vector of strings)}

\item{value_var}{dependent variable (string)}

\item{between}{between-subject variables (vector of strings)}

\item{adjustments}{adjustment settings as used for \link[superb]{superData} (default: single CI estimates, Cousineau-Morey adjustment)}
}
\value{
dataframe with computed CIs
}
\description{
Get within-subject CI using the superb package
}
\examples{
data(faces)
get_superb_ci(faces, 'uid', 'stim_gender', 'answerTime')

}
