% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{Arrange several plots with one legend and/or one x-axis title}
\usage{
grid_arrange_shared_legend(..., stack = "v", one_sub = F, heights = F,
  one_x_axis = F, legend_pos = "b")
}
\arguments{
\item{...}{}

\item{stack}{}

\item{one_sub}{}

\item{heights}{}
}
\value{
plots arranged with grid.arrange
}
\description{
Based on a function from https://github.com/hadley/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs
}

