% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{lmer.fixef}
\alias{lmer.fixef}
\title{Get nice matrix of fixed effects from lmer}
\usage{
lmer.fixef(fit.lmer)
}
\arguments{
\item{fit.lmer}{}
}
\value{
result
}
\description{
Get nice matrix of fixed effects from lmer
}

