% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lmer_with_julia}
\alias{lmer_with_julia}
\title{Run lmer with Julia}
\usage{
lmer_with_julia(myform, dataset)
}
\arguments{
\item{dataset}{}
}
\value{
result
}
\description{
Fits maximal model
}
\examples{

load.libs(c('lme4','rjulia','data.table'))
julia_init()
data(faces)

ptm <- proc.time()
summary(lmer(answerTime~correct*stim_gender+(1+correct*stim_gender|uid), faces))
proc.time() - ptm

ptm <- proc.time()
lmer_with_julia(answerTime~correct*stim_gender+(1|uid), faces)
proc.time() - ptm

#Bizzare model - user gender cannot be a random effect within users, only for demonstration
ptm <- proc.time()
summary(lmer(answerTime~correct*stim_gender*user_gender+(1+correct*stim_gender*user_gender|uid), faces))
proc.time() - ptm

ptm <- proc.time()
lmer_with_julia(answerTime~correct*stim_gender*user_gender+(1|uid), faces)
proc.time() - ptm

}

