% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{paste_and}
\alias{paste_and}
\title{Paste several strings, add 'and' before last}
\usage{
paste_and(x, sep = ", ", suffix = "")
}
\arguments{
\item{suffix}{}
}
\value{
result
}
\description{
Paste several strings, add 'and' before last
}

