% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.pointrange}
\alias{plot.pointrange}
\title{Title}
\usage{
\method{generic}{class}(..., pos = position_dodge(0.3), pointsize = I(3),
  linesize = I(1), pointfill = I("white"), within_subj = F, wid = "uid",
  bars = "ci", withinvars = NULL, betweenvars = NULL, x_as_numeric = F,
  add_h_line = NULL, connecting_line = F, pretty_breaks_y = F,
  pretty_y_axis = F, exp_y = F, print_aggregated_data = F)
}
\arguments{
\item{...}{}

\item{pos}{}

\item{pointsize}{}

\item{linesize}{}

\item{pointfill}{}

\item{within_subj}{}

\item{wid}{}

\item{bars}{}

\item{withinvars}{}

\item{betweenvars}{}

\item{x_as_numeric}{}

\item{add_h_line}{}

\item{connecting_line}{}
}
\value{
plot of pointrange
}
\description{
Title
}

