% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.pointrange}
\alias{plot.pointrange}
\title{Pointrange plot}
\usage{
\method{generic}{class}(
  ...,
  pos = position_dodge(0.3),
  pointsize = I(3),
  linesize = I(1),
  pointfill = I("white"),
  pointshape = NULL,
  within_subj = F,
  wid = "uid",
  bars = "ci",
  withinvars = NULL,
  betweenvars = NULL,
  x_as_numeric = F,
  custom_geom_before = NULL,
  connecting_line = F,
  pretty_breaks_y = F,
  pretty_y_axis = F,
  exp_y = F,
  print_aggregated_data = F,
  do_aggregate = F,
  add_margin = F,
  margin_label = "all",
  margin_x_vals = NULL,
  bars_instead_of_points = F,
  geom_bar_params = NULL
)
}
\arguments{
\item{...}{the first element is assumed to be the dataset and the second one is assumed to be a list of plot aesthetics (e.g., aes(x = varA, y=varB))}

\item{pos}{position adjustment function (e.g., position_dodge())}

\item{pointsize}{size for the points (e.g., "I(1)")}

\item{linesize}{size for the lines (e.g., "I(1)")}

\item{pointfill}{fill for the points (e.g., "I('white')")}

\item{pointshape}{shape for the points (e.g., "I(22)")}

\item{within_subj}{should we use within-subject adjustment?}

\item{wid}{within-subject ID variable}

\item{bars}{should we use confidence intervals ("ci") or standard errors ("se")?}

\item{withinvars}{within-subject variables}

\item{betweenvars}{between-subject variables}

\item{x_as_numeric}{should we treat x as numeric (otherwise it is treated as is)?}

\item{custom_geom_before}{any custom geom to add before points and connecting line (geoms that should be on top of the pointrange can be added in a usual manner)}

\item{connecting_line}{should we add connecting line (T/F)?}

\item{pretty_breaks_y}{prettify y breaks}

\item{pretty_y_axis}{prettify y axis}

\item{exp_y}{exponentiate y (it's better to use this than scale_y_exp if pretty_breaks_y is used)}

\item{print_aggregated_data}{print aggregated data used for plotting to console}

\item{do_aggregate}{aggregate data by all conditions before plotting (False)}

\item{add_margin}{add margin that would show the aggregate over all or subset of x-axis values (T/F)}

\item{margin_label}{margin label to use}

\item{margin_x_vals}{which levels of x-axis variable to aggregate over (NULL means all of them)}

\item{bars_instead_of_points}{use geom_bar instead of geom_point}

\item{geom_bar_params}{parameters for geom_bar if it is used}
}
\value{
plot of pointrange
}
\description{
Pointrange plot
}
\details{
For point and line properties (e.g., pointfill) passing NULL allows to avoid setting these values (useful when they are mapped to some variables).
}
\examples{
data(faces)
plot.pointrange(faces, aes(x=user_gender, color=stim_gender, y=answerTime))+ylab('RT')
plot.pointrange(faces, aes(x=user_gender, color=stim_gender, y=answerTime), wid='uid', within_subj=T, withinvars=c('stim_gender'), betweenvars=c('user_gender'))+ylab('RT')
plot.pointrange(faces, aes(x=user_gender, shape=stim_gender, y=answerTime), wid='uid', withinvars=c('stim_gender'), betweenvars=c('user_gender'), within_subj=T, bars='se')+ylab('RT')
plot.pointrange(faces, aes(x=user_gender, shape=stim_gender, y=answerTime), wid='uid', withinvars=c('stim_gender'), betweenvars=c('user_gender'), within_subj=T, bars='se', print_aggregated_data=T)+ylab('RT')
plot.pointrange(faces, aes(x=user_gender, shape=stim_gender, y=answerTime), withinvars=c('stim_gender'), betweenvars = c('user_gender'), print_aggregated_data = T, wid='uid', within_subj=T, exp_y=T, bars='ci', do_aggregate = T)+ylab('RT')
}
