% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.pointrange}
\alias{plot.pointrange}
\title{Pointrange plot}
\usage{
\method{generic}{class}(..., pos = position_dodge(0.3), pointsize = I(3),
  linesize = I(1), pointfill = I("white"), within_subj = F, wid = "uid",
  bars = "ci", withinvars = NULL, betweenvars = NULL, x_as_numeric = F,
  add_h_line = NULL, connecting_line = F, pretty_breaks_y = F,
  pretty_y_axis = F, exp_y = F, print_aggregated_data = F,
  do_aggregate = F)
}
\arguments{
\item{...}{}

\item{pos}{}

\item{pointsize}{}

\item{linesize}{}

\item{pointfill}{}

\item{within_subj}{}

\item{wid}{}

\item{bars}{}

\item{withinvars}{}

\item{betweenvars}{}

\item{x_as_numeric}{}

\item{add_h_line}{}

\item{connecting_line}{}

\item{pretty_breaks_y}{}

\item{pretty_y_axis}{}

\item{exp_y}{}

\item{print_aggregated_data}{}

\item{do_aggregate}{- aggregate data by all conditions before plotting (False)}
}
\value{
plot of pointrange
}
\description{
Pointrange plot
}
\examples{
data(faces)
plot.pointrange(faces, aes(x=user_gender, color=stim_gender, y=answerTime))+ylab('RT')
plot.pointrange(faces, aes(x=user_gender, color=stim_gender, y=answerTime), wid='uid', within_subj=T)+ylab('RT')
plot.pointrange(faces, aes(x=user_gender, shape=stim_gender, y=answerTime), wid='uid', within_subj=T, bars='se')+ylab('RT')
plot.pointrange(faces, aes(x=user_gender, shape=stim_gender, y=answerTime), wid='uid', within_subj=T, bars='se', print_aggregated_data=T)+ylab('RT')
}

