% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{round.p}
\alias{round.p}
\title{Round *p*-value}
\usage{
\method{generic}{class #'}(list, include.rel = 1, digits = 3)
}
\arguments{
\item{list}{a vector of p-values}

\item{include.rel}{include relation sign}

\item{digits}{a number of decimal digits}
}
\value{
Formatted p-value
}
\description{
If p-value is <= 0.001, returns ".001" else returns p-value rounded to the specified number of digits, optionally including relation sign ("<" or "=").
}
\examples{
round.p(c(0.025, 0.0001, 0.001, 0.568))
round.p(c(0.025, 0.0001, 0.001, 0.568), digits=2)
round.p(c(0.025, 0.0001, 0.001, 0.568), include.rel=F)
}

