% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scale_y_exp}
\alias{scale_y_exp}
\title{Transform exponential scale back to normal}
\usage{
scale_y_exp(digits = 0, ...)
}
\arguments{
\item{...}{other arguments passed to scale_y_continuous}
}
\value{
scale
}
\description{
Transform exponential scale back to normal
}

