% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{table.mean.conf.by}
\alias{table.mean.conf.by}
\title{Show table of means and confidence intervals by group}
\usage{
table.mean.conf.by(x, by, digits = 2, binom = F)
}
\arguments{
\item{binom}{}
}
\value{
result
}
\description{
Show table of means and confidence intervals by group
}
