% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{table.mean.conf}
\alias{table.mean.conf}
\title{Get a list with means and confidence intervals}
\usage{
table.mean.conf(x, digits = 2, binom = FALSE, ...)
}
\arguments{
\item{x}{value to compute the mean for}

\item{digits}{number of digits in results (default: 2)}

\item{binom}{compute binomial CI instead of the usual ones}

\item{...}{other parameters passed to \link{format.results}}
}
\value{
a list with a mean and CI
}
\description{
Get a list with means and confidence intervals
}
\examples{
table.mean.conf(faces$correct)

}
