% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{table_mean_conf}
\alias{table_mean_conf}
\title{Get a list with means and confidence intervals}
\usage{
table_mean_conf(x, digits = 2, binom = FALSE, ...)
}
\arguments{
\item{x}{numeric vector to compute the mean for}

\item{digits}{number of digits in results (default: 2)}

\item{binom}{compute binomial CI instead of the usual ones}

\item{...}{other parameters passed to \link{format.results}}
}
\value{
a list with a mean and CI as formatted strings
}
\description{
Get a list with means and confidence intervals
}
\examples{
# For continuous data
x <- rnorm(100, mean = 50, sd = 10)
table_mean_conf(x)
table_mean_conf(x, digits = 3)

# For binary data
y <- rbinom(100, 1, prob = 0.7)
table_mean_conf(y, binom = TRUE)
}
